Fix equipment lists to their own module for export
