+++ src/equipment.js
@@ -1,0 +1,6554 @@
+
+(function(self) {
+
+  const handItems = [{
+    index: 1, 
+    itemName: "Monster vial 1", 
+    ramAddr: 0x0A4B38, 
+    isoAddr: 0x0B1110, 
+    giveItemR4: 0x01, 
+    giveItemR5: 0x00, 
+    atkVal: 40, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B111C, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x08, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 20, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0099, 
+    menuPal: 0x0097, 
+    critRate: 1
+  }, {
+    index: 2, 
+    itemName: "Monster vial 2", 
+    ramAddr: 0x0A4B6C, 
+    isoAddr: 0x0B1144, 
+    giveItemR4: 0x02, 
+    giveItemR5: 0x00, 
+    atkVal: 25, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1150, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x08, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 20, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0099, 
+    menuPal: 0x0093, 
+    critRate: 1
+  }, {
+    index: 3, 
+    itemName: "Monster vial 3", 
+    ramAddr: 0x0A4BA0, 
+    isoAddr: 0x0B1178, 
+    giveItemR4: 0x03, 
+    giveItemR5: 0x00, 
+    atkVal: 30, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1184, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x08, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 20, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0099, 
+    menuPal: 0x0096, 
+    critRate: 1
+  }, {
+    index: 4, 
+    itemName: "Shield rod", 
+    ramAddr: 0x0A4BD4, 
+    isoAddr: 0x0B11AC, 
+    giveItemR4: 0x04, 
+    giveItemR5: 0x00, 
+    atkVal: 11, 
+    defVal: 2, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B11B8, 
+    categoryTxt: "Rod", 
+    categoryHex: 0x04, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 20, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0005, 
+    menuPal: 0x0005, 
+    critRate: 1
+  }, {
+    index: 5, 
+    itemName: "Leather shield", 
+    ramAddr: 0x0A4C08, 
+    isoAddr: 0x0B11E0, 
+    giveItemR4: 0x05, 
+    giveItemR5: 0x00, 
+    atkVal: 1, 
+    defVal: 1, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B11EC, 
+    categoryTxt: "Shield", 
+    categoryHex: 0x09, 
+    lockDur: 0, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 6, 
+    shieldSpell: 0x00000002, 
+    mpUse: 0, 
+    hitType: 0x0004, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0006, 
+    menuPal: 0x0006, 
+    critRate: 1
+  }, {
+    index: 6, 
+    itemName: "Knight shield", 
+    ramAddr: 0x0A4C3C, 
+    isoAddr: 0x0B1214, 
+    giveItemR4: 0x06, 
+    giveItemR5: 0x00, 
+    atkVal: 1, 
+    defVal: 2, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1220, 
+    categoryTxt: "Shield", 
+    categoryHex: 0x09, 
+    lockDur: 0, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 6, 
+    shieldSpell: 0x00000004, 
+    mpUse: 0, 
+    hitType: 0x0004, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0007, 
+    menuPal: 0x0007, 
+    critRate: 1
+  }, {
+    index: 7, 
+    itemName: "Iron shield", 
+    ramAddr: 0x0A4C70, 
+    isoAddr: 0x0B1248, 
+    giveItemR4: 0x07, 
+    giveItemR5: 0x00, 
+    atkVal: 1, 
+    defVal: 2, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1254, 
+    categoryTxt: "Shield", 
+    categoryHex: 0x09, 
+    lockDur: 0, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 6, 
+    shieldSpell: 0x00000008, 
+    mpUse: 0, 
+    hitType: 0x0004, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0008, 
+    menuPal: 0x0008, 
+    critRate: 1
+  }, {
+    index: 8, 
+    itemName: "AxeLord shield", 
+    ramAddr: 0x0A4CA4, 
+    isoAddr: 0x0B127C, 
+    giveItemR4: 0x08, 
+    giveItemR5: 0x00, 
+    atkVal: 1, 
+    defVal: 3, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1288, 
+    categoryTxt: "Shield", 
+    categoryHex: 0x09, 
+    lockDur: 0, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 6, 
+    shieldSpell: 0x00000010, 
+    mpUse: 0, 
+    hitType: 0x0004, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0009, 
+    menuPal: 0x0009, 
+    critRate: 1
+  }, {
+    index: 9, 
+    itemName: "Herald Shield", 
+    ramAddr: 0x0A4CD8, 
+    isoAddr: 0x0B12B0, 
+    giveItemR4: 0x09, 
+    giveItemR5: 0x00, 
+    atkVal: 1, 
+    defVal: 4, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B12BC, 
+    categoryTxt: "Shield", 
+    categoryHex: 0x09, 
+    lockDur: 0, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 6, 
+    shieldSpell: 0x00000020, 
+    mpUse: 0, 
+    hitType: 0x0004, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x000A, 
+    menuPal: 0x000A, 
+    critRate: 1
+  }, {
+    index: 10, 
+    itemName: "Dark shield", 
+    ramAddr: 0x0A4D0C, 
+    isoAddr: 0x0B12E4, 
+    giveItemR4: 0x0A, 
+    giveItemR5: 0x00, 
+    atkVal: 10, 
+    defVal: 3, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B12F0, 
+    categoryTxt: "Shield", 
+    categoryHex: 0x09, 
+    lockDur: 0, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 30, 
+    shieldSpell: 0x00000040, 
+    mpUse: 0, 
+    hitType: 0x0006, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x000B, 
+    menuPal: 0x000B, 
+    critRate: 1
+  }, {
+    index: 11, 
+    itemName: "Goddess shield", 
+    ramAddr: 0x0A4D40, 
+    isoAddr: 0x0B1318, 
+    giveItemR4: 0x0B, 
+    giveItemR5: 0x00, 
+    atkVal: 1, 
+    defVal: 4, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1324, 
+    categoryTxt: "Shield", 
+    categoryHex: 0x09, 
+    lockDur: 0, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 6, 
+    shieldSpell: 0x00000080, 
+    mpUse: 0, 
+    hitType: 0x0004, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x000C, 
+    menuPal: 0x000C, 
+    critRate: 1
+  }, {
+    index: 12, 
+    itemName: "Shaman shield", 
+    ramAddr: 0x0A4D74, 
+    isoAddr: 0x0B134C, 
+    giveItemR4: 0x0C, 
+    giveItemR5: 0x00, 
+    atkVal: 1, 
+    defVal: 4, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1358, 
+    categoryTxt: "Shield", 
+    categoryHex: 0x09, 
+    lockDur: 0, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 6, 
+    shieldSpell: 0x00000100, 
+    mpUse: 0, 
+    hitType: 0x0004, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x000D, 
+    menuPal: 0x000D, 
+    critRate: 1
+  }, {
+    index: 13, 
+    itemName: "Medusa shield", 
+    ramAddr: 0x0A4DA8, 
+    isoAddr: 0x0B1380, 
+    giveItemR4: 0x0D, 
+    giveItemR5: 0x00, 
+    atkVal: 13, 
+    defVal: 1, 
+    elementTxt: "STN", 
+    elementHex: 0x0200, 
+    elementOffset: 0x0B138C, 
+    categoryTxt: "Shield", 
+    categoryHex: 0x09, 
+    lockDur: 0, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 30, 
+    shieldSpell: 0x00000200, 
+    mpUse: 0, 
+    hitType: 0x0006, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x000E, 
+    menuPal: 0x000E, 
+    critRate: 1
+  }, {
+    index: 14, 
+    itemName: "Skull shield", 
+    ramAddr: 0x0A4DDC, 
+    isoAddr: 0x0B13B4, 
+    giveItemR4: 0x0E, 
+    giveItemR5: 0x00, 
+    atkVal: 1, 
+    defVal: 4, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B13C0, 
+    categoryTxt: "Shield", 
+    categoryHex: 0x09, 
+    lockDur: 0, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 6, 
+    shieldSpell: 0x00000400, 
+    mpUse: 0, 
+    hitType: 0x0004, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x000F, 
+    menuPal: 0x000F, 
+    critRate: 1
+  }, {
+    index: 15, 
+    itemName: "Fire shield", 
+    ramAddr: 0x0A4E10, 
+    isoAddr: 0x0B13E8, 
+    giveItemR4: 0x0F, 
+    giveItemR5: 0x00, 
+    atkVal: 1, 
+    defVal: 4, 
+    elementTxt: "FIR", 
+    elementHex: 0x8000, 
+    elementOffset: 0x0B13F4, 
+    categoryTxt: "Shield", 
+    categoryHex: 0x09, 
+    lockDur: 0, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 6, 
+    shieldSpell: 0x00000800, 
+    mpUse: 0, 
+    hitType: 0x0004, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0010, 
+    menuPal: 0x0010, 
+    critRate: 1
+  }, {
+    index: 16, 
+    itemName: "Alucard shield", 
+    ramAddr: 0x0A4E44, 
+    isoAddr: 0x0B141C, 
+    giveItemR4: 0x10, 
+    giveItemR5: 0x00, 
+    atkVal: 20, 
+    defVal: 5, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1428, 
+    categoryTxt: "Shield", 
+    categoryHex: 0x09, 
+    lockDur: 0, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 6, 
+    shieldSpell: 0x00001000, 
+    mpUse: 0, 
+    hitType: 0x0004, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0004, 
+    menuPal: 0x0004, 
+    critRate: 1
+  }, {
+    index: 17, 
+    itemName: "Sword of Dawn", 
+    ramAddr: 0x0A4E78, 
+    isoAddr: 0x0B1450, 
+    giveItemR4: 0x11, 
+    giveItemR5: 0x00, 
+    atkVal: 27, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B145C, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0xB1, 
+    useItem: 0x00, 
+    nmeIFrame: 16, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0011, 
+    menuPal: 0x0011, 
+    critRate: 1
+  }, {
+    index: 18, 
+    itemName: "Basilard", 
+    ramAddr: 0x0A4EAC, 
+    isoAddr: 0x0B1484, 
+    giveItemR4: 0x12, 
+    giveItemR5: 0x00, 
+    atkVal: 1, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B1490, 
+    categoryTxt: "ShortSword", 
+    categoryHex: 0x00, 
+    lockDur: 5, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0012, 
+    menuPal: 0x0012, 
+    critRate: 1
+  }, {
+    index: 19, 
+    itemName: "Short sword", 
+    ramAddr: 0x0A4EE0, 
+    isoAddr: 0x0B14B8, 
+    giveItemR4: 0x13, 
+    giveItemR5: 0x00, 
+    atkVal: 2, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B14C4, 
+    categoryTxt: "ShortSword", 
+    categoryHex: 0x00, 
+    lockDur: 8, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0013, 
+    menuPal: 0x0013, 
+    critRate: 1
+  }, {
+    index: 20, 
+    itemName: "Combat knife", 
+    ramAddr: 0x0A4F14, 
+    isoAddr: 0x0B14EC, 
+    giveItemR4: 0x14, 
+    giveItemR5: 0x00, 
+    atkVal: 24, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B14F8, 
+    categoryTxt: "ShortSword", 
+    categoryHex: 0x00, 
+    lockDur: 5, 
+    chainLimit: 0x80, 
+    special: 0xAF, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0014, 
+    menuPal: 0x0014, 
+    critRate: 32
+  }, {
+    index: 21, 
+    itemName: "Nunchaku", 
+    ramAddr: 0x0A4F48, 
+    isoAddr: 0x0B1520, 
+    giveItemR4: 0x15, 
+    giveItemR5: 0x00, 
+    atkVal: 10, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B152C, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 16, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0015, 
+    menuPal: 0x0015, 
+    critRate: 32
+  }, {
+    index: 22, 
+    itemName: "Were Bane", 
+    ramAddr: 0x0A4F7C, 
+    isoAddr: 0x0B1554, 
+    giveItemR4: 0x16, 
+    giveItemR5: 0x00, 
+    atkVal: 18, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B1560, 
+    categoryTxt: "ShortSword", 
+    categoryHex: 0x00, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0xBE, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0016, 
+    menuPal: 0x0016, 
+    critRate: 32
+  }, {
+    index: 23, 
+    itemName: "Rapier", 
+    ramAddr: 0x0A4FB0, 
+    isoAddr: 0x0B1588, 
+    giveItemR4: 0x17, 
+    giveItemR5: 0x00, 
+    atkVal: 7, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B1594, 
+    categoryTxt: "ShortSword", 
+    categoryHex: 0x00, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0xBF, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0017, 
+    menuPal: 0x0017, 
+    critRate: 32
+  }, {
+    index: 24, 
+    itemName: "Karma Coin", 
+    ramAddr: 0x0A4FE4, 
+    isoAddr: 0x0B15BC, 
+    giveItemR4: 0x18, 
+    giveItemR5: 0x00, 
+    atkVal: 50, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B15C8, 
+    categoryTxt: "Bomb", 
+    categoryHex: 0x07, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0018, 
+    menuPal: 0x0018, 
+    critRate: 1
+  }, {
+    index: 25, 
+    itemName: "Magic Missile", 
+    ramAddr: 0x0A5018, 
+    isoAddr: 0x0B1720, 
+    giveItemR4: 0x19, 
+    giveItemR5: 0x00, 
+    atkVal: 30, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B172C, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x04, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0019, 
+    menuPal: 0x0019, 
+    critRate: 1
+  }, {
+    index: 26, 
+    itemName: "Red Rust", 
+    ramAddr: 0x0A504C, 
+    isoAddr: 0x0B1754, 
+    giveItemR4: 0x1A, 
+    giveItemR5: 0x00, 
+    atkVal: -2, 
+    defVal: 0, 
+    elementTxt: "CUR+CUT", 
+    elementHex: 0x0140, 
+    elementOffset: 0x0B1760, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x001A, 
+    menuPal: 0x001A, 
+    critRate: 32
+  }, {
+    index: 27, 
+    itemName: "Takemitsu", 
+    ramAddr: 0x0A5080, 
+    isoAddr: 0x0B1788, 
+    giveItemR4: 0x1B, 
+    giveItemR5: 0x00, 
+    atkVal: 1, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1794, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x001B, 
+    menuPal: 0x001B, 
+    critRate: 32
+  }, {
+    index: 28, 
+    itemName: "Shotel", 
+    ramAddr: 0x0A50B4, 
+    isoAddr: 0x0B17BC, 
+    giveItemR4: 0x1C, 
+    giveItemR5: 0x00, 
+    atkVal: 20, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B17C8, 
+    categoryTxt: "ShortSword", 
+    categoryHex: 0x00, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0xB0, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x001C, 
+    menuPal: 0x001C, 
+    critRate: 1
+  }, {
+    index: 29, 
+    itemName: "Orange", 
+    ramAddr: 0x0A50E8, 
+    isoAddr: 0x0B17F0, 
+    giveItemR4: 0x1D, 
+    giveItemR5: 0x00, 
+    atkVal: 6, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B17FC, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x001D, 
+    menuPal: 0x001D, 
+    critRate: 1
+  }, {
+    index: 30, 
+    itemName: "Apple", 
+    ramAddr: 0x0A511C, 
+    isoAddr: 0x0B1824, 
+    giveItemR4: 0x1E, 
+    giveItemR5: 0x00, 
+    atkVal: 9, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1830, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x001E, 
+    menuPal: 0x001E, 
+    critRate: 1
+  }, {
+    index: 31, 
+    itemName: "Banana", 
+    ramAddr: 0x0A5150, 
+    isoAddr: 0x0B1858, 
+    giveItemR4: 0x1F, 
+    giveItemR5: 0x00, 
+    atkVal: 7, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1864, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x001F, 
+    menuPal: 0x001F, 
+    critRate: 1
+  }, {
+    index: 32, 
+    itemName: "Grapes", 
+    ramAddr: 0x0A5184, 
+    isoAddr: 0x0B188C, 
+    giveItemR4: 0x20, 
+    giveItemR5: 0x00, 
+    atkVal: 5, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1898, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0020, 
+    menuPal: 0x0020, 
+    critRate: 1
+  }, {
+    index: 33, 
+    itemName: "Strawberry", 
+    ramAddr: 0x0A51B8, 
+    isoAddr: 0x0B18C0, 
+    giveItemR4: 0x21, 
+    giveItemR5: 0x00, 
+    atkVal: 10, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B18CC, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0021, 
+    menuPal: 0x0021, 
+    critRate: 1
+  }, {
+    index: 34, 
+    itemName: "Pineapple", 
+    ramAddr: 0x0A51EC, 
+    isoAddr: 0x0B18F4, 
+    giveItemR4: 0x22, 
+    giveItemR5: 0x00, 
+    atkVal: 8, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1900, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0022, 
+    menuPal: 0x0022, 
+    critRate: 1
+  }, {
+    index: 35, 
+    itemName: "Peanuts", 
+    ramAddr: 0x0A5220, 
+    isoAddr: 0x0B1928, 
+    giveItemR4: 0x23, 
+    giveItemR5: 0x00, 
+    atkVal: 50, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1934, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0023, 
+    menuPal: 0x0023, 
+    critRate: 1
+  }, {
+    index: 36, 
+    itemName: "Toadstool", 
+    ramAddr: 0x0A5254, 
+    isoAddr: 0x0B195C, 
+    giveItemR4: 0x24, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1968, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0024, 
+    menuPal: 0x0024, 
+    critRate: 1
+  }, {
+    index: 37, 
+    itemName: "Shiitake", 
+    ramAddr: 0x0A5288, 
+    isoAddr: 0x0B1990, 
+    giveItemR4: 0x25, 
+    giveItemR5: 0x00, 
+    atkVal: 30, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B199C, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0025, 
+    menuPal: 0x0025, 
+    critRate: 1
+  }, {
+    index: 38, 
+    itemName: "Cheesecake", 
+    ramAddr: 0x0A52BC, 
+    isoAddr: 0x0B19C4, 
+    giveItemR4: 0x26, 
+    giveItemR5: 0x00, 
+    atkVal: 14, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B19D0, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0026, 
+    menuPal: 0x0026, 
+    critRate: 1
+  }, {
+    index: 39, 
+    itemName: "Shortcake", 
+    ramAddr: 0x0A52F0, 
+    isoAddr: 0x0B19F8, 
+    giveItemR4: 0x27, 
+    giveItemR5: 0x00, 
+    atkVal: 11, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1A04, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0027, 
+    menuPal: 0x0027, 
+    critRate: 1
+  }, {
+    index: 40, 
+    itemName: "Tart", 
+    ramAddr: 0x0A5324, 
+    isoAddr: 0x0B1A2C, 
+    giveItemR4: 0x28, 
+    giveItemR5: 0x00, 
+    atkVal: 12, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1A38, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0028, 
+    menuPal: 0x0028, 
+    critRate: 1
+  }, {
+    index: 41, 
+    itemName: "Parfait", 
+    ramAddr: 0x0A5358, 
+    isoAddr: 0x0B1A60, 
+    giveItemR4: 0x29, 
+    giveItemR5: 0x00, 
+    atkVal: 16, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1A6C, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0029, 
+    menuPal: 0x0029, 
+    critRate: 1
+  }, {
+    index: 42, 
+    itemName: "Pudding", 
+    ramAddr: 0x0A538C, 
+    isoAddr: 0x0B1A94, 
+    giveItemR4: 0x2A, 
+    giveItemR5: 0x00, 
+    atkVal: 17, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1AA0, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x002A, 
+    menuPal: 0x002A, 
+    critRate: 1
+  }, {
+    index: 43, 
+    itemName: "Ice cream", 
+    ramAddr: 0x0A53C0, 
+    isoAddr: 0x0B1AC8, 
+    giveItemR4: 0x2B, 
+    giveItemR5: 0x00, 
+    atkVal: 15, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1AD4, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x002B, 
+    menuPal: 0x002B, 
+    critRate: 1
+  }, {
+    index: 44, 
+    itemName: "Frankfurter", 
+    ramAddr: 0x0A53F4, 
+    isoAddr: 0x0B1AFC, 
+    giveItemR4: 0x2C, 
+    giveItemR5: 0x00, 
+    atkVal: 19, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1B08, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x002C, 
+    menuPal: 0x002C, 
+    critRate: 1
+  }, {
+    index: 45, 
+    itemName: "Hamburger", 
+    ramAddr: 0x0A5428, 
+    isoAddr: 0x0B1B30, 
+    giveItemR4: 0x2D, 
+    giveItemR5: 0x00, 
+    atkVal: 20, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1B3C, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x002D, 
+    menuPal: 0x002D, 
+    critRate: 1
+  }, {
+    index: 46, 
+    itemName: "Pizza", 
+    ramAddr: 0x0A545C, 
+    isoAddr: 0x0B1B64, 
+    giveItemR4: 0x2E, 
+    giveItemR5: 0x00, 
+    atkVal: 21, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1B70, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x002E, 
+    menuPal: 0x002E, 
+    critRate: 1
+  }, {
+    index: 47, 
+    itemName: "Cheese", 
+    ramAddr: 0x0A5490, 
+    isoAddr: 0x0B1B98, 
+    giveItemR4: 0x2F, 
+    giveItemR5: 0x00, 
+    atkVal: 18, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1BA4, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x002F, 
+    menuPal: 0x002F, 
+    critRate: 1
+  }, {
+    index: 48, 
+    itemName: "Ham and eggs", 
+    ramAddr: 0x0A54C4, 
+    isoAddr: 0x0B1BCC, 
+    giveItemR4: 0x30, 
+    giveItemR5: 0x00, 
+    atkVal: 22, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1BD8, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0030, 
+    menuPal: 0x0030, 
+    critRate: 1
+  }, {
+    index: 49, 
+    itemName: "Omelette", 
+    ramAddr: 0x0A54F8, 
+    isoAddr: 0x0B1C00, 
+    giveItemR4: 0x31, 
+    giveItemR5: 0x00, 
+    atkVal: 23, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1C0C, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0031, 
+    menuPal: 0x0031, 
+    critRate: 1
+  }, {
+    index: 50, 
+    itemName: "Morning set", 
+    ramAddr: 0x0A552C, 
+    isoAddr: 0x0B1C34, 
+    giveItemR4: 0x32, 
+    giveItemR5: 0x00, 
+    atkVal: 24, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1C40, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0032, 
+    menuPal: 0x0032, 
+    critRate: 1
+  }, {
+    index: 51, 
+    itemName: "Lunch A", 
+    ramAddr: 0x0A5560, 
+    isoAddr: 0x0B1C68, 
+    giveItemR4: 0x33, 
+    giveItemR5: 0x00, 
+    atkVal: 25, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1C74, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0033, 
+    menuPal: 0x0033, 
+    critRate: 1
+  }, {
+    index: 52, 
+    itemName: "Lunch B", 
+    ramAddr: 0x0A5594, 
+    isoAddr: 0x0B1C9C, 
+    giveItemR4: 0x34, 
+    giveItemR5: 0x00, 
+    atkVal: 26, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1CA8, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0034, 
+    menuPal: 0x0034, 
+    critRate: 1
+  }, {
+    index: 53, 
+    itemName: "Curry rice", 
+    ramAddr: 0x0A55C8, 
+    isoAddr: 0x0B1CD0, 
+    giveItemR4: 0x35, 
+    giveItemR5: 0x00, 
+    atkVal: 28, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1CDC, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0035, 
+    menuPal: 0x0035, 
+    critRate: 1
+  }, {
+    index: 54, 
+    itemName: "Gyros plate", 
+    ramAddr: 0x0A55FC, 
+    isoAddr: 0x0B1D04, 
+    giveItemR4: 0x36, 
+    giveItemR5: 0x00, 
+    atkVal: 29, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1D10, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0036, 
+    menuPal: 0x0036, 
+    critRate: 1
+  }, {
+    index: 55, 
+    itemName: "Spaghetti", 
+    ramAddr: 0x0A5630, 
+    isoAddr: 0x0B1D38, 
+    giveItemR4: 0x37, 
+    giveItemR5: 0x00, 
+    atkVal: 27, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1D44, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0037, 
+    menuPal: 0x0037, 
+    critRate: 1
+  }, {
+    index: 56, 
+    itemName: "Grape juice", 
+    ramAddr: 0x0A5664, 
+    isoAddr: 0x0B1D6C, 
+    giveItemR4: 0x38, 
+    giveItemR5: 0x00, 
+    atkVal: 20, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1D78, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0038, 
+    menuPal: 0x0038, 
+    critRate: 1
+  }, {
+    index: 57, 
+    itemName: "Barley tea", 
+    ramAddr: 0x0A5698, 
+    isoAddr: 0x0B1DA0, 
+    giveItemR4: 0x39, 
+    giveItemR5: 0x00, 
+    atkVal: 15, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1DAC, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0039, 
+    menuPal: 0x0039, 
+    critRate: 1
+  }, {
+    index: 58, 
+    itemName: "Green tea", 
+    ramAddr: 0x0A56CC, 
+    isoAddr: 0x0B1DD4, 
+    giveItemR4: 0x3A, 
+    giveItemR5: 0x00, 
+    atkVal: 10, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1DE0, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x003A, 
+    menuPal: 0x003A, 
+    critRate: 1
+  }, {
+    index: 59, 
+    itemName: "Natou", 
+    ramAddr: 0x0A5700, 
+    isoAddr: 0x0B1E08, 
+    giveItemR4: 0x3B, 
+    giveItemR5: 0x00, 
+    atkVal: 32, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1E14, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x003B, 
+    menuPal: 0x003B, 
+    critRate: 1
+  }, {
+    index: 60, 
+    itemName: "Ramen", 
+    ramAddr: 0x0A5734, 
+    isoAddr: 0x0B1E3C, 
+    giveItemR4: 0x3C, 
+    giveItemR5: 0x00, 
+    atkVal: 30, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1E48, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x003C, 
+    menuPal: 0x003C, 
+    critRate: 1
+  }, {
+    index: 61, 
+    itemName: "Miso soup", 
+    ramAddr: 0x0A5768, 
+    isoAddr: 0x0B1E70, 
+    giveItemR4: 0x3D, 
+    giveItemR5: 0x00, 
+    atkVal: 20, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1E7C, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x003D, 
+    menuPal: 0x003D, 
+    critRate: 1
+  }, {
+    index: 62, 
+    itemName: "Sushi", 
+    ramAddr: 0x0A579C, 
+    isoAddr: 0x0B1EA4, 
+    giveItemR4: 0x3E, 
+    giveItemR5: 0x00, 
+    atkVal: 100, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1EB0, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x003E, 
+    menuPal: 0x003E, 
+    critRate: 1
+  }, {
+    index: 63, 
+    itemName: "Pork bun", 
+    ramAddr: 0x0A57D0, 
+    isoAddr: 0x0B1ED8, 
+    giveItemR4: 0x3F, 
+    giveItemR5: 0x00, 
+    atkVal: 17, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B1EE4, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x003F, 
+    menuPal: 0x003F, 
+    critRate: 1
+  }, {
+    index: 64, 
+    itemName: "Red bean bun", 
+    ramAddr: 0x0A5804, 
+    isoAddr: 0x0B203C, 
+    giveItemR4: 0x40, 
+    giveItemR5: 0x00, 
+    atkVal: 12, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B2048, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0040, 
+    menuPal: 0x0040, 
+    critRate: 1
+  }, {
+    index: 65, 
+    itemName: "Chinese bun", 
+    ramAddr: 0x0A5838, 
+    isoAddr: 0x0B2070, 
+    giveItemR4: 0x41, 
+    giveItemR5: 0x00, 
+    atkVal: 10, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B207C, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0041, 
+    menuPal: 0x0041, 
+    critRate: 1
+  }, {
+    index: 66, 
+    itemName: "Dim Sum set", 
+    ramAddr: 0x0A586C, 
+    isoAddr: 0x0B20A4, 
+    giveItemR4: 0x42, 
+    giveItemR5: 0x00, 
+    atkVal: 30, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B20B0, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0042, 
+    menuPal: 0x0042, 
+    critRate: 1
+  }, {
+    index: 67, 
+    itemName: "Pot roast", 
+    ramAddr: 0x0A58A0, 
+    isoAddr: 0x0B20D8, 
+    giveItemR4: 0x43, 
+    giveItemR5: 0x00, 
+    atkVal: 50, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B20E4, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0043, 
+    menuPal: 0x0043, 
+    critRate: 1
+  }, {
+    index: 68, 
+    itemName: "Sirloin", 
+    ramAddr: 0x0A58D4, 
+    isoAddr: 0x0B210C, 
+    giveItemR4: 0x44, 
+    giveItemR5: 0x00, 
+    atkVal: 100, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B2118, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0044, 
+    menuPal: 0x0044, 
+    critRate: 1
+  }, {
+    index: 69, 
+    itemName: "Turkey", 
+    ramAddr: 0x0A5908, 
+    isoAddr: 0x0B2140, 
+    giveItemR4: 0x45, 
+    giveItemR5: 0x00, 
+    atkVal: 80, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B214C, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0045, 
+    menuPal: 0x0045, 
+    critRate: 1
+  }, {
+    index: 70, 
+    itemName: "Meal ticket", 
+    ramAddr: 0x0A593C, 
+    isoAddr: 0x0B2174, 
+    giveItemR4: 0x46, 
+    giveItemR5: 0x00, 
+    atkVal: 40, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B2180, 
+    categoryTxt: "Food", 
+    categoryHex: 0x06, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0046, 
+    menuPal: 0x0046, 
+    critRate: 1
+  }, {
+    index: 71, 
+    itemName: "Neutron bomb", 
+    ramAddr: 0x0A5970, 
+    isoAddr: 0x0B21A8, 
+    giveItemR4: 0x47, 
+    giveItemR5: 0x00, 
+    atkVal: 30, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B21B4, 
+    categoryTxt: "Bomb", 
+    categoryHex: 0x07, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 30, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0082, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0047, 
+    menuPal: 0x0047, 
+    critRate: 1
+  }, {
+    index: 72, 
+    itemName: "Power of Sire", 
+    ramAddr: 0x0A59A4, 
+    isoAddr: 0x0B21DC, 
+    giveItemR4: 0x48, 
+    giveItemR5: 0x00, 
+    atkVal: 40, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B21E8, 
+    categoryTxt: "Bomb", 
+    categoryHex: 0x07, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 8, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0082, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0048, 
+    menuPal: 0x0048, 
+    critRate: 1
+  }, {
+    index: 73, 
+    itemName: "Pentagram", 
+    ramAddr: 0x0A59D8, 
+    isoAddr: 0x0B2210, 
+    giveItemR4: 0x49, 
+    giveItemR5: 0x00, 
+    atkVal: 50, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B221C, 
+    categoryTxt: "Bomb", 
+    categoryHex: 0x07, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 30, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0082, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0049, 
+    menuPal: 0x0049, 
+    critRate: 1
+  }, {
+    index: 74, 
+    itemName: "Bat Pentagram", 
+    ramAddr: 0x0A5A0C, 
+    isoAddr: 0x0B2244, 
+    giveItemR4: 0x4A, 
+    giveItemR5: 0x00, 
+    atkVal: 99, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B2250, 
+    categoryTxt: "Bomb", 
+    categoryHex: 0x07, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 30, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0082, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x004A, 
+    menuPal: 0x004A, 
+    critRate: 1
+  }, {
+    index: 75, 
+    itemName: "Shuriken", 
+    ramAddr: 0x0A5A40, 
+    isoAddr: 0x0B2278, 
+    giveItemR4: 0x4B, 
+    giveItemR5: 0x00, 
+    atkVal: 50, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2284, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x03, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x004B, 
+    menuPal: 0x004B, 
+    critRate: 32
+  }, {
+    index: 76, 
+    itemName: "Cross shuriken", 
+    ramAddr: 0x0A5A74, 
+    isoAddr: 0x0B22AC, 
+    giveItemR4: 0x4C, 
+    giveItemR5: 0x00, 
+    atkVal: 70, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B22B8, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x03, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x004C, 
+    menuPal: 0x004C, 
+    critRate: 32
+  }, {
+    index: 77, 
+    itemName: "Buffalo star", 
+    ramAddr: 0x0A5AA8, 
+    isoAddr: 0x0B22E0, 
+    giveItemR4: 0x4D, 
+    giveItemR5: 0x00, 
+    atkVal: 90, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B22EC, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x03, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x004D, 
+    menuPal: 0x004D, 
+    critRate: 32
+  }, {
+    index: 78, 
+    itemName: "Flame star", 
+    ramAddr: 0x0A5ADC, 
+    isoAddr: 0x0B2314, 
+    giveItemR4: 0x4E, 
+    giveItemR5: 0x00, 
+    atkVal: 100, 
+    defVal: 0, 
+    elementTxt: "FIR+CUT", 
+    elementHex: 0x8040, 
+    elementOffset: 0x0B2320, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x03, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x004E, 
+    menuPal: 0x004E, 
+    critRate: 32
+  }, {
+    index: 79, 
+    itemName: "TNT", 
+    ramAddr: 0x0A5B10, 
+    isoAddr: 0x0B2348, 
+    giveItemR4: 0x4F, 
+    giveItemR5: 0x00, 
+    atkVal: 60, 
+    defVal: 0, 
+    elementTxt: "FIR", 
+    elementHex: 0x8000, 
+    elementOffset: 0x0B2354, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 30, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x004F, 
+    menuPal: 0x004F, 
+    critRate: 1
+  }, {
+    index: 80, 
+    itemName: "Bwaka knife", 
+    ramAddr: 0x0A5B44, 
+    isoAddr: 0x0B237C, 
+    giveItemR4: 0x50, 
+    giveItemR5: 0x00, 
+    atkVal: 30, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2388, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x03, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0050, 
+    menuPal: 0x0050, 
+    critRate: 1
+  }, {
+    index: 81, 
+    itemName: "Boomerang", 
+    ramAddr: 0x0A5B78, 
+    isoAddr: 0x0B23B0, 
+    giveItemR4: 0x51, 
+    giveItemR5: 0x00, 
+    atkVal: 30, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B23BC, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x03, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0051, 
+    menuPal: 0x0051, 
+    critRate: 1
+  }, {
+    index: 82, 
+    itemName: "Javelin", 
+    ramAddr: 0x0A5BAC, 
+    isoAddr: 0x0B23E4, 
+    giveItemR4: 0x52, 
+    giveItemR5: 0x00, 
+    atkVal: 50, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B23F0, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0052, 
+    menuPal: 0x0052, 
+    critRate: 1
+  }, {
+    index: 83, 
+    itemName: "Tyrfing", 
+    ramAddr: 0x0A5BE0, 
+    isoAddr: 0x0B2418, 
+    giveItemR4: 0x53, 
+    giveItemR5: 0x00, 
+    atkVal: -30, 
+    defVal: 0, 
+    elementTxt: "DRK+CUR+CUT", 
+    elementHex: 0x0940, 
+    elementOffset: 0x0B2424, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 18, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0053, 
+    menuPal: 0x0053, 
+    critRate: 1
+  }, {
+    index: 84, 
+    itemName: "Namakura", 
+    ramAddr: 0x0A5C14, 
+    isoAddr: 0x0B244C, 
+    giveItemR4: 0x54, 
+    giveItemR5: 0x00, 
+    atkVal: -5, 
+    defVal: 0, 
+    elementTxt: "CUR+CUT", 
+    elementHex: 0x0140, 
+    elementOffset: 0x0B2458, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 18, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0054, 
+    menuPal: 0x0054, 
+    critRate: 1
+  }, {
+    index: 85, 
+    itemName: "Knuckle duster", 
+    ramAddr: 0x0A5C48, 
+    isoAddr: 0x0B2480, 
+    giveItemR4: 0x55, 
+    giveItemR5: 0x00, 
+    atkVal: 1, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B248C, 
+    categoryTxt: "Fist", 
+    categoryHex: 0x03, 
+    lockDur: 5, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0055, 
+    menuPal: 0x0055, 
+    critRate: 64
+  }, {
+    index: 86, 
+    itemName: "Gladius", 
+    ramAddr: 0x0A5C7C, 
+    isoAddr: 0x0B24B4, 
+    giveItemR4: 0x56, 
+    giveItemR5: 0x00, 
+    atkVal: 3, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B24C0, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0056, 
+    menuPal: 0x0056, 
+    critRate: 1
+  }, {
+    index: 87, 
+    itemName: "Scimitar", 
+    ramAddr: 0x0A5CB0, 
+    isoAddr: 0x0B24E8, 
+    giveItemR4: 0x57, 
+    giveItemR5: 0x00, 
+    atkVal: 5, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B24F4, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0057, 
+    menuPal: 0x0057, 
+    critRate: 1
+  }, {
+    index: 88, 
+    itemName: "Cutlass", 
+    ramAddr: 0x0A5CE4, 
+    isoAddr: 0x0B251C, 
+    giveItemR4: 0x58, 
+    giveItemR5: 0x00, 
+    atkVal: 7, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2528, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0058, 
+    menuPal: 0x0058, 
+    critRate: 1
+  }, {
+    index: 89, 
+    itemName: "Saber", 
+    ramAddr: 0x0A5D18, 
+    isoAddr: 0x0B2550, 
+    giveItemR4: 0x59, 
+    giveItemR5: 0x00, 
+    atkVal: 9, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B255C, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0059, 
+    menuPal: 0x0059, 
+    critRate: 1
+  }, {
+    index: 90, 
+    itemName: "Falchion", 
+    ramAddr: 0x0A5D4C, 
+    isoAddr: 0x0B2584, 
+    giveItemR4: 0x5A, 
+    giveItemR5: 0x00, 
+    atkVal: 10, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2590, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x005A, 
+    menuPal: 0x005A, 
+    critRate: 1
+  }, {
+    index: 91, 
+    itemName: "Broadsword", 
+    ramAddr: 0x0A5D80, 
+    isoAddr: 0x0B25B8, 
+    giveItemR4: 0x5B, 
+    giveItemR5: 0x00, 
+    atkVal: 12, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B25C4, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x005B, 
+    menuPal: 0x005B, 
+    critRate: 1
+  }, {
+    index: 92, 
+    itemName: "Bekatowa", 
+    ramAddr: 0x0A5DB4, 
+    isoAddr: 0x0B25EC, 
+    giveItemR4: 0x5C, 
+    giveItemR5: 0x00, 
+    atkVal: 14, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B25F8, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x005C, 
+    menuPal: 0x005C, 
+    critRate: 1
+  }, {
+    index: 93, 
+    itemName: "Damascus sword", 
+    ramAddr: 0x0A5DE8, 
+    isoAddr: 0x0B2620, 
+    giveItemR4: 0x5D, 
+    giveItemR5: 0x00, 
+    atkVal: 17, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B262C, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x005D, 
+    menuPal: 0x005D, 
+    critRate: 1
+  }, {
+    index: 94, 
+    itemName: "Hunter sword", 
+    ramAddr: 0x0A5E1C, 
+    isoAddr: 0x0B2654, 
+    giveItemR4: 0x5E, 
+    giveItemR5: 0x00, 
+    atkVal: 20, 
+    defVal: -1, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2660, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x005E, 
+    menuPal: 0x005E, 
+    critRate: 8
+  }, {
+    index: 95, 
+    itemName: "Estoc", 
+    ramAddr: 0x0A5E50, 
+    isoAddr: 0x0B2688, 
+    giveItemR4: 0x5F, 
+    giveItemR5: 0x00, 
+    atkVal: 39, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2694, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 18, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 28, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x005F, 
+    menuPal: 0x005F, 
+    critRate: 32
+  }, {
+    index: 96, 
+    itemName: "Bastard sword", 
+    ramAddr: 0x0A5E84, 
+    isoAddr: 0x0B26BC, 
+    giveItemR4: 0x60, 
+    giveItemR5: 0x00, 
+    atkVal: 20, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B26C8, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0060, 
+    menuPal: 0x0060, 
+    critRate: 1
+  }, {
+    index: 97, 
+    itemName: "Jewel knuckles", 
+    ramAddr: 0x0A5EB8, 
+    isoAddr: 0x0B26F0, 
+    giveItemR4: 0x61, 
+    giveItemR5: 0x00, 
+    atkVal: 21, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B26FC, 
+    categoryTxt: "Fist", 
+    categoryHex: 0x03, 
+    lockDur: 5, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0061, 
+    menuPal: 0x0061, 
+    critRate: 64
+  }, {
+    index: 98, 
+    itemName: "Claymore", 
+    ramAddr: 0x0A5EEC, 
+    isoAddr: 0x0B2724, 
+    giveItemR4: 0x62, 
+    giveItemR5: 0x00, 
+    atkVal: 26, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2730, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 15, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 15, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0062, 
+    menuPal: 0x0062, 
+    critRate: 1
+  }, {
+    index: 99, 
+    itemName: "Talwar", 
+    ramAddr: 0x0A5F20, 
+    isoAddr: 0x0B2758, 
+    giveItemR4: 0x63, 
+    giveItemR5: 0x00, 
+    atkVal: 22, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2764, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0063, 
+    menuPal: 0x0063, 
+    critRate: 1
+  }, {
+    index: 100, 
+    itemName: "Katana", 
+    ramAddr: 0x0A5F54, 
+    isoAddr: 0x0B278C, 
+    giveItemR4: 0x64, 
+    giveItemR5: 0x00, 
+    atkVal: 32, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2798, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0xBD, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0064, 
+    menuPal: 0x0064, 
+    critRate: 32
+  }, {
+    index: 101, 
+    itemName: "Flamberge", 
+    ramAddr: 0x0A5F88, 
+    isoAddr: 0x0B27C0, 
+    giveItemR4: 0x65, 
+    giveItemR5: 0x00, 
+    atkVal: 31, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B27CC, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 15, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 15, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0065, 
+    menuPal: 0x0065, 
+    critRate: 1
+  }, {
+    index: 102, 
+    itemName: "Iron Fist", 
+    ramAddr: 0x0A5FBC, 
+    isoAddr: 0x0B27F4, 
+    giveItemR4: 0x66, 
+    giveItemR5: 0x00, 
+    atkVal: 26, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B2800, 
+    categoryTxt: "Fist", 
+    categoryHex: 0x03, 
+    lockDur: 5, 
+    chainLimit: 0x80, 
+    special: 0xBA, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0066, 
+    menuPal: 0x0066, 
+    critRate: 64
+  }, {
+    index: 103, 
+    itemName: "Zwei hander", 
+    ramAddr: 0x0A5FF0, 
+    isoAddr: 0x0B2828, 
+    giveItemR4: 0x67, 
+    giveItemR5: 0x00, 
+    atkVal: 38, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2834, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 15, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 15, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0067, 
+    menuPal: 0x0067, 
+    critRate: 1
+  }, {
+    index: 104, 
+    itemName: "Sword of Hador", 
+    ramAddr: 0x0A6024, 
+    isoAddr: 0x0B298C, 
+    giveItemR4: 0x68, 
+    giveItemR5: 0x00, 
+    atkVal: 24, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2998, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0068, 
+    menuPal: 0x0068, 
+    critRate: 1
+  }, {
+    index: 105, 
+    itemName: "Luminus", 
+    ramAddr: 0x0A6058, 
+    isoAddr: 0x0B29C0, 
+    giveItemR4: 0x69, 
+    giveItemR5: 0x00, 
+    atkVal: 26, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B29CC, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0069, 
+    menuPal: 0x0069, 
+    critRate: 1
+  }, {
+    index: 106, 
+    itemName: "Harper", 
+    ramAddr: 0x0A608C, 
+    isoAddr: 0x0B29F4, 
+    giveItemR4: 0x6A, 
+    giveItemR5: 0x00, 
+    atkVal: 28, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2A00, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x006A, 
+    menuPal: 0x006A, 
+    critRate: 1
+  }, {
+    index: 107, 
+    itemName: "Obsidian sword", 
+    ramAddr: 0x0A60C0, 
+    isoAddr: 0x0B2A28, 
+    giveItemR4: 0x6B, 
+    giveItemR5: 0x00, 
+    atkVal: 42, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2A34, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 15, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x006B, 
+    menuPal: 0x006B, 
+    critRate: 1
+  }, {
+    index: 108, 
+    itemName: "Gram", 
+    ramAddr: 0x0A60F4, 
+    isoAddr: 0x0B2A5C, 
+    giveItemR4: 0x6C, 
+    giveItemR5: 0x00, 
+    atkVal: 30, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2A68, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x006C, 
+    menuPal: 0x006C, 
+    critRate: 1
+  }, {
+    index: 109, 
+    itemName: "Jewel sword", 
+    ramAddr: 0x0A6128, 
+    isoAddr: 0x0B2A90, 
+    giveItemR4: 0x6D, 
+    giveItemR5: 0x00, 
+    atkVal: 15, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2A9C, 
+    categoryTxt: "ShortSword", 
+    categoryHex: 0x00, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0xC6, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0085, 
+    menuIcon: 0x006D, 
+    menuPal: 0x006D, 
+    critRate: 1
+  }, {
+    index: 110, 
+    itemName: "Mormegil", 
+    ramAddr: 0x0A615C, 
+    isoAddr: 0x0B2AC4, 
+    giveItemR4: 0x6E, 
+    giveItemR5: 0x00, 
+    atkVal: 25, 
+    defVal: 0, 
+    elementTxt: "DRK+CUT", 
+    elementHex: 0x0840, 
+    elementOffset: 0x0B2AD0, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x006E, 
+    menuPal: 0x006E, 
+    critRate: 1
+  }, {
+    index: 111, 
+    itemName: "Firebrand", 
+    ramAddr: 0x0A6190, 
+    isoAddr: 0x0B2AF8, 
+    giveItemR4: 0x6F, 
+    giveItemR5: 0x00, 
+    atkVal: 25, 
+    defVal: 0, 
+    elementTxt: "FIR+CUT", 
+    elementHex: 0x8040, 
+    elementOffset: 0x0B2B04, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0xB3, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x006F, 
+    menuPal: 0x006F, 
+    critRate: 1
+  }, {
+    index: 112, 
+    itemName: "Thunderbrand", 
+    ramAddr: 0x0A61C4, 
+    isoAddr: 0x0B2B2C, 
+    giveItemR4: 0x70, 
+    giveItemR5: 0x00, 
+    atkVal: 25, 
+    defVal: 0, 
+    elementTxt: "LIT+CUT", 
+    elementHex: 0x4040, 
+    elementOffset: 0x0B2B38, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0xB2, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0070, 
+    menuPal: 0x0070, 
+    critRate: 1
+  }, {
+    index: 113, 
+    itemName: "Icebrand", 
+    ramAddr: 0x0A61F8, 
+    isoAddr: 0x0B2B60, 
+    giveItemR4: 0x71, 
+    giveItemR5: 0x00, 
+    atkVal: 25, 
+    defVal: 0, 
+    elementTxt: "ICE+CUT", 
+    elementHex: 0x2040, 
+    elementOffset: 0x0B2B6C, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0xB5, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0071, 
+    menuPal: 0x0071, 
+    critRate: 1
+  }, {
+    index: 114, 
+    itemName: "Stone Sword", 
+    ramAddr: 0x0A622C, 
+    isoAddr: 0x0B2B94, 
+    giveItemR4: 0x72, 
+    giveItemR5: 0x00, 
+    atkVal: 1, 
+    defVal: 0, 
+    elementTxt: "STN", 
+    elementHex: 0x0200, 
+    elementOffset: 0x0B2BA0, 
+    categoryTxt: "ShortSword", 
+    categoryHex: 0x00, 
+    lockDur: 8, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0072, 
+    menuPal: 0x0072, 
+    critRate: 1
+  }, {
+    index: 115, 
+    itemName: "Holy Sword", 
+    ramAddr: 0x0A6260, 
+    isoAddr: 0x0B2BC8, 
+    giveItemR4: 0x73, 
+    giveItemR5: 0x00, 
+    atkVal: 26, 
+    defVal: 0, 
+    elementTxt: "HLY+CUT", 
+    elementHex: 0x1040, 
+    elementOffset: 0x0B2BD4, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0xC4, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0073, 
+    menuPal: 0x0073, 
+    critRate: 1
+  }, {
+    index: 116, 
+    itemName: "Terminus Est", 
+    ramAddr: 0x0A6294, 
+    isoAddr: 0x0B2BFC, 
+    giveItemR4: 0x74, 
+    giveItemR5: 0x00, 
+    atkVal: 32, 
+    defVal: 2, 
+    elementTxt: "PSN+CUT", 
+    elementHex: 0x00C0, 
+    elementOffset: 0x0B2C08, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0074, 
+    menuPal: 0x0074, 
+    critRate: 1
+  }, {
+    index: 117, 
+    itemName: "Marsil", 
+    ramAddr: 0x0A62C8, 
+    isoAddr: 0x0B2C30, 
+    giveItemR4: 0x75, 
+    giveItemR5: 0x00, 
+    atkVal: 33, 
+    defVal: 1, 
+    elementTxt: "FIR+CUT", 
+    elementHex: 0x8040, 
+    elementOffset: 0x0B2C3C, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0xB4, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0075, 
+    menuPal: 0x0075, 
+    critRate: 1
+  }, {
+    index: 118, 
+    itemName: "Dark Blade", 
+    ramAddr: 0x0A62FC, 
+    isoAddr: 0x0B2C64, 
+    giveItemR4: 0x76, 
+    giveItemR5: 0x00, 
+    atkVal: 35, 
+    defVal: 2, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2C70, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0076, 
+    menuPal: 0x0076, 
+    critRate: 1
+  }, {
+    index: 119, 
+    itemName: "Heaven Sword", 
+    ramAddr: 0x0A6330, 
+    isoAddr: 0x0B2C98, 
+    giveItemR4: 0x77, 
+    giveItemR5: 0x00, 
+    atkVal: 21, 
+    defVal: -2, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2CA4, 
+    categoryTxt: "Throw1", 
+    categoryHex: 0x02, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 20, 
+    shieldSpell: 0x00000001, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0077, 
+    menuPal: 0x0077, 
+    critRate: 32
+  }, {
+    index: 120, 
+    itemName: "Fist of Tulkas", 
+    ramAddr: 0x0A6364, 
+    isoAddr: 0x0B2CCC, 
+    giveItemR4: 0x78, 
+    giveItemR5: 0x00, 
+    atkVal: 38, 
+    defVal: 5, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B2CD8, 
+    categoryTxt: "Fist", 
+    categoryHex: 0x03, 
+    lockDur: 5, 
+    chainLimit: 0x80, 
+    special: 0xBC, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0078, 
+    menuPal: 0x0078, 
+    critRate: 64
+  }, {
+    index: 121, 
+    itemName: "Gurthang", 
+    ramAddr: 0x0A6398, 
+    isoAddr: 0x0B2D00, 
+    giveItemR4: 0x79, 
+    giveItemR5: 0x00, 
+    atkVal: 25, 
+    defVal: 1, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2D0C, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0xC3, 
+    useItem: 0x00, 
+    nmeIFrame: 12, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0079, 
+    menuPal: 0x0079, 
+    critRate: 1
+  }, {
+    index: 122, 
+    itemName: "Mourneblade", 
+    ramAddr: 0x0A63CC, 
+    isoAddr: 0x0B2D34, 
+    giveItemR4: 0x7A, 
+    giveItemR5: 0x00, 
+    atkVal: 36, 
+    defVal: 1, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2D40, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0082, 
+    menuIcon: 0x007A, 
+    menuPal: 0x007A, 
+    critRate: 1
+  }, {
+    index: 123, 
+    itemName: "Alucard sword", 
+    ramAddr: 0x0A6400, 
+    isoAddr: 0x0B2D68, 
+    giveItemR4: 0x7B, 
+    giveItemR5: 0x00, 
+    atkVal: 42, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2D74, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 10, 
+    chainLimit: 0x80, 
+    special: 0xC5, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x007B, 
+    menuPal: 0x007B, 
+    critRate: 1
+  }, {
+    index: 124, 
+    itemName: "Mablung Sword", 
+    ramAddr: 0x0A6434, 
+    isoAddr: 0x0B2D9C, 
+    giveItemR4: 0x7C, 
+    giveItemR5: 0x00, 
+    atkVal: 39, 
+    defVal: 2, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2DA8, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x007C, 
+    menuPal: 0x007C, 
+    critRate: 1
+  }, {
+    index: 125, 
+    itemName: "Badelaire", 
+    ramAddr: 0x0A6468, 
+    isoAddr: 0x0B2DD0, 
+    giveItemR4: 0x7D, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2DDC, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x007D, 
+    menuPal: 0x007D, 
+    critRate: 1
+  }, {
+    index: 126, 
+    itemName: "Sword Familiar", 
+    ramAddr: 0x0A649C, 
+    isoAddr: 0x0B2E04, 
+    giveItemR4: 0x7E, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2E10, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x007E, 
+    menuPal: 0x007E, 
+    critRate: 1
+  }, {
+    index: 127, 
+    itemName: "Great Sword", 
+    ramAddr: 0x0A64D0, 
+    isoAddr: 0x0B2E38, 
+    giveItemR4: 0x7F, 
+    giveItemR5: 0x00, 
+    atkVal: 65, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2E44, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 18, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 15, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x007F, 
+    menuPal: 0x007F, 
+    critRate: 1
+  }, {
+    index: 128, 
+    itemName: "Mace", 
+    ramAddr: 0x0A6504, 
+    isoAddr: 0x0B2E6C, 
+    giveItemR4: 0x80, 
+    giveItemR5: 0x00, 
+    atkVal: 9, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B2E78, 
+    categoryTxt: "Rod", 
+    categoryHex: 0x04, 
+    lockDur: 11, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 20, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0080, 
+    menuPal: 0x0080, 
+    critRate: 1
+  }, {
+    index: 129, 
+    itemName: "Morningstar", 
+    ramAddr: 0x0A6538, 
+    isoAddr: 0x0B2EA0, 
+    giveItemR4: 0x81, 
+    giveItemR5: 0x00, 
+    atkVal: 11, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B2EAC, 
+    categoryTxt: "Rod", 
+    categoryHex: 0x04, 
+    lockDur: 11, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 22, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0081, 
+    menuPal: 0x0081, 
+    critRate: 1
+  }, {
+    index: 130, 
+    itemName: "Holy Rod", 
+    ramAddr: 0x0A656C, 
+    isoAddr: 0x0B2ED4, 
+    giveItemR4: 0x82, 
+    giveItemR5: 0x00, 
+    atkVal: 12, 
+    defVal: 0, 
+    elementTxt: "HLY", 
+    elementHex: 0x1000, 
+    elementOffset: 0x0B2EE0, 
+    categoryTxt: "Rod", 
+    categoryHex: 0x04, 
+    lockDur: 11, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 20, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0082, 
+    menuPal: 0x0082, 
+    critRate: 1
+  }, {
+    index: 131, 
+    itemName: "Star Flail", 
+    ramAddr: 0x0A65A0, 
+    isoAddr: 0x0B2F08, 
+    giveItemR4: 0x83, 
+    giveItemR5: 0x00, 
+    atkVal: 15, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B2F14, 
+    categoryTxt: "Rod", 
+    categoryHex: 0x04, 
+    lockDur: 11, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 20, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0083, 
+    menuPal: 0x0083, 
+    critRate: 1
+  }, {
+    index: 132, 
+    itemName: "Moon Rod", 
+    ramAddr: 0x0A65D4, 
+    isoAddr: 0x0B2F3C, 
+    giveItemR4: 0x84, 
+    giveItemR5: 0x00, 
+    atkVal: 17, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B2F48, 
+    categoryTxt: "Rod", 
+    categoryHex: 0x04, 
+    lockDur: 11, 
+    chainLimit: 0x80, 
+    special: 0xAC, 
+    useItem: 0x00, 
+    nmeIFrame: 20, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0084, 
+    menuPal: 0x0084, 
+    critRate: 1
+  }, {
+    index: 133, 
+    itemName: "Chakram", 
+    ramAddr: 0x0A6608, 
+    isoAddr: 0x0B2F70, 
+    giveItemR4: 0x85, 
+    giveItemR5: 0x00, 
+    atkVal: 15, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B2F7C, 
+    categoryTxt: "Throw1", 
+    categoryHex: 0x02, 
+    lockDur: 0, 
+    chainLimit: 0x02, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 30, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0085, 
+    menuPal: 0x0085, 
+    critRate: 32
+  }, {
+    index: 134, 
+    itemName: "Fire boomerang", 
+    ramAddr: 0x0A663C, 
+    isoAddr: 0x0B2FA4, 
+    giveItemR4: 0x86, 
+    giveItemR5: 0x00, 
+    atkVal: 50, 
+    defVal: 0, 
+    elementTxt: "FIR", 
+    elementHex: 0x8000, 
+    elementOffset: 0x0B2FB0, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x03, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 30, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0086, 
+    menuPal: 0x0086, 
+    critRate: 1
+  }, {
+    index: 135, 
+    itemName: "Iron Ball", 
+    ramAddr: 0x0A6670, 
+    isoAddr: 0x0B2FD8, 
+    giveItemR4: 0x87, 
+    giveItemR5: 0x00, 
+    atkVal: 60, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B2FE4, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x04, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 30, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0087, 
+    menuPal: 0x0087, 
+    critRate: 1
+  }, {
+    index: 136, 
+    itemName: "Holbein dagger", 
+    ramAddr: 0x0A66A4, 
+    isoAddr: 0x0B300C, 
+    giveItemR4: 0x88, 
+    giveItemR5: 0x00, 
+    atkVal: 6, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B3018, 
+    categoryTxt: "ShortSword", 
+    categoryHex: 0x00, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0012, 
+    menuPal: 0x0012, 
+    critRate: 32
+  }, {
+    index: 137, 
+    itemName: "Blue knuckle", 
+    ramAddr: 0x0A66D8, 
+    isoAddr: 0x0B3040, 
+    giveItemR4: 0x89, 
+    giveItemR5: 0x00, 
+    atkVal: 8, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B304C, 
+    categoryTxt: "Fist", 
+    categoryHex: 0x03, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 18, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0055, 
+    menuPal: 0x0055, 
+    critRate: 64
+  }, {
+    index: 138, 
+    itemName: "Dynamite", 
+    ramAddr: 0x0A670C, 
+    isoAddr: 0x0B3074, 
+    giveItemR4: 0x8A, 
+    giveItemR5: 0x00, 
+    atkVal: 100, 
+    defVal: 0, 
+    elementTxt: "FIR", 
+    elementHex: 0x8000, 
+    elementOffset: 0x0B3080, 
+    categoryTxt: "Throw2", 
+    categoryHex: 0x08, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 16, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0088, 
+    menuPal: 0x004F, 
+    critRate: 1
+  }, {
+    index: 139, 
+    itemName: "Osafune katana", 
+    ramAddr: 0x0A6740, 
+    isoAddr: 0x0B30A8, 
+    giveItemR4: 0x8B, 
+    giveItemR5: 0x00, 
+    atkVal: 50, 
+    defVal: -5, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B30B4, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0xC1, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0064, 
+    menuPal: 0x0064, 
+    critRate: 32
+  }, {
+    index: 140, 
+    itemName: "Masamune", 
+    ramAddr: 0x0A6774, 
+    isoAddr: 0x0B30DC, 
+    giveItemR4: 0x8C, 
+    giveItemR5: 0x00, 
+    atkVal: 90, 
+    defVal: -15, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B30E8, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0xC2, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0064, 
+    menuPal: 0x0064, 
+    critRate: 64
+  }, {
+    index: 141, 
+    itemName: "Muramasa", 
+    ramAddr: 0x0A67A8, 
+    isoAddr: 0x0B3110, 
+    giveItemR4: 0x8D, 
+    giveItemR5: 0x00, 
+    atkVal: -5, 
+    defVal: -2, 
+    elementTxt: "CUR+CUT", 
+    elementHex: 0x0140, 
+    elementOffset: 0x0B311C, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 13, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x001A, 
+    menuPal: 0x001A, 
+    critRate: 1
+  }, {
+    index: 142, 
+    itemName: "Heart Refresh", 
+    ramAddr: 0x0A67DC, 
+    isoAddr: 0x0B3144, 
+    giveItemR4: 0x8E, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B3150, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x00, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x008E, 
+    menuPal: 0x008E, 
+    critRate: 1
+  }, {
+    index: 143, 
+    itemName: "Runesword", 
+    ramAddr: 0x0A6810, 
+    isoAddr: 0x0B32A8, 
+    giveItemR4: 0x8F, 
+    giveItemR5: 0x00, 
+    atkVal: 40, 
+    defVal: -4, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B32B4, 
+    categoryTxt: "Throw1", 
+    categoryHex: 0x02, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 20, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x008F, 
+    menuPal: 0x008F, 
+    critRate: 32
+  }, {
+    index: 144, 
+    itemName: "Antivenom", 
+    ramAddr: 0x0A6844, 
+    isoAddr: 0x0B32DC, 
+    giveItemR4: 0x90, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B32E8, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0090, 
+    menuPal: 0x0090, 
+    critRate: 1
+  }, {
+    index: 145, 
+    itemName: "Uncurse", 
+    ramAddr: 0x0A6878, 
+    isoAddr: 0x0B3310, 
+    giveItemR4: 0x91, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B331C, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0091, 
+    menuPal: 0x0090, 
+    critRate: 1
+  }, {
+    index: 146, 
+    itemName: "Life apple", 
+    ramAddr: 0x0A68AC, 
+    isoAddr: 0x0B3344, 
+    giveItemR4: 0x92, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B3350, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x009A, 
+    menuPal: 0x0095, 
+    critRate: 1
+  }, {
+    index: 147, 
+    itemName: "Hammer", 
+    ramAddr: 0x0A68E0, 
+    isoAddr: 0x0B3378, 
+    giveItemR4: 0x93, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B3384, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x009B, 
+    menuPal: 0x0095, 
+    critRate: 1
+  }, {
+    index: 148, 
+    itemName: "Str. potion", 
+    ramAddr: 0x0A6914, 
+    isoAddr: 0x0B33AC, 
+    giveItemR4: 0x94, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B33B8, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0095, 
+    menuPal: 0x0095, 
+    critRate: 1
+  }, {
+    index: 149, 
+    itemName: "Luck potion", 
+    ramAddr: 0x0A6948, 
+    isoAddr: 0x0B33E0, 
+    giveItemR4: 0x95, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B33EC, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0096, 
+    menuPal: 0x0097, 
+    critRate: 1
+  }, {
+    index: 150, 
+    itemName: "Smart potion", 
+    ramAddr: 0x0A697C, 
+    isoAddr: 0x0B3414, 
+    giveItemR4: 0x96, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B3420, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0096, 
+    menuPal: 0x0095, 
+    critRate: 1
+  }, {
+    index: 151, 
+    itemName: "Attack potion", 
+    ramAddr: 0x0A69B0, 
+    isoAddr: 0x0B3448, 
+    giveItemR4: 0x97, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B3454, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0096, 
+    menuPal: 0x0093, 
+    critRate: 1
+  }, {
+    index: 152, 
+    itemName: "Shield potion", 
+    ramAddr: 0x0A69E4, 
+    isoAddr: 0x0B347C, 
+    giveItemR4: 0x98, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B3488, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0096, 
+    menuPal: 0x0096, 
+    critRate: 1
+  }, {
+    index: 153, 
+    itemName: "Resist fire", 
+    ramAddr: 0x0A6A18, 
+    isoAddr: 0x0B34B0, 
+    giveItemR4: 0x99, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B34BC, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0094, 
+    menuPal: 0x0094, 
+    critRate: 1
+  }, {
+    index: 154, 
+    itemName: "Resist thunder", 
+    ramAddr: 0x0A6A4C, 
+    isoAddr: 0x0B34E4, 
+    giveItemR4: 0x9A, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B34F0, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0094, 
+    menuPal: 0x0096, 
+    critRate: 1
+  }, {
+    index: 155, 
+    itemName: "Resist ice", 
+    ramAddr: 0x0A6A80, 
+    isoAddr: 0x0B3518, 
+    giveItemR4: 0x9B, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B3524, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0094, 
+    menuPal: 0x0098, 
+    critRate: 1
+  }, {
+    index: 156, 
+    itemName: "Resist stone", 
+    ramAddr: 0x0A6AB4, 
+    isoAddr: 0x0B354C, 
+    giveItemR4: 0x9C, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B3558, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0094, 
+    menuPal: 0x0091, 
+    critRate: 1
+  }, {
+    index: 157, 
+    itemName: "Resist holy", 
+    ramAddr: 0x0A6AE8, 
+    isoAddr: 0x0B3580, 
+    giveItemR4: 0x9D, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B358C, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0094, 
+    menuPal: 0x0095, 
+    critRate: 1
+  }, {
+    index: 158, 
+    itemName: "Resist dark", 
+    ramAddr: 0x0A6B1C, 
+    isoAddr: 0x0B35B4, 
+    giveItemR4: 0x9E, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B35C0, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0094, 
+    menuPal: 0x009A, 
+    critRate: 1
+  }, {
+    index: 159, 
+    itemName: "Potion", 
+    ramAddr: 0x0A6B50, 
+    isoAddr: 0x0B35E8, 
+    giveItemR4: 0x9F, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B35F4, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0092, 
+    menuPal: 0x0097, 
+    critRate: 1
+  }, {
+    index: 160, 
+    itemName: "High potion", 
+    ramAddr: 0x0A6B84, 
+    isoAddr: 0x0B361C, 
+    giveItemR4: 0xA0, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B3628, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0093, 
+    menuPal: 0x0097, 
+    critRate: 1
+  }, {
+    index: 161, 
+    itemName: "Elixir", 
+    ramAddr: 0x0A6BB8, 
+    isoAddr: 0x0B3650, 
+    giveItemR4: 0xA1, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B365C, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0097, 
+    menuPal: 0x0097, 
+    critRate: 1
+  }, {
+    index: 162, 
+    itemName: "Manna prism", 
+    ramAddr: 0x0A6BEC, 
+    isoAddr: 0x0B3684, 
+    giveItemR4: 0xA2, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B3690, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0098, 
+    menuPal: 0x0099, 
+    critRate: 1
+  }, {
+    index: 163, 
+    itemName: "Vorpal blade", 
+    ramAddr: 0x0A6C20, 
+    isoAddr: 0x0B36B8, 
+    giveItemR4: 0xA3, 
+    giveItemR5: 0x00, 
+    atkVal: 25, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B36C4, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 0, 
+    chainLimit: 0x03, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0077, 
+    menuPal: 0x0077, 
+    critRate: 1
+  }, {
+    index: 164, 
+    itemName: "Crissaegrim", 
+    ramAddr: 0x0A6C54, 
+    isoAddr: 0x0B36EC, 
+    giveItemR4: 0xA4, 
+    giveItemR5: 0x00, 
+    atkVal: 36, 
+    defVal: 1, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B36F8, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 0, 
+    chainLimit: 0x03, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0077, 
+    menuPal: 0x0077, 
+    critRate: 1
+  }, {
+    index: 165, 
+    itemName: "Yasutsuna", 
+    ramAddr: 0x0A6C88, 
+    isoAddr: 0x0B3720, 
+    giveItemR4: 0xA5, 
+    giveItemR5: 0x00, 
+    atkVal: 60, 
+    defVal: -10, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B372C, 
+    categoryTxt: "2-Hand", 
+    categoryHex: 0x05, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 10, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x0064, 
+    menuPal: 0x0064, 
+    critRate: 64
+  }, {
+    index: 166, 
+    itemName: "Library card", 
+    ramAddr: 0x0A6CBC, 
+    isoAddr: 0x0B3754, 
+    giveItemR4: 0xA6, 
+    giveItemR5: 0x00, 
+    atkVal: 0, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B3760, 
+    categoryTxt: "Medicine", 
+    categoryHex: 0x0A, 
+    lockDur: 0, 
+    chainLimit: 0x01, 
+    special: 0x00, 
+    useItem: 0x01, 
+    nmeIFrame: 0, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0000, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0046, 
+    menuPal: 0x0046, 
+    critRate: 1
+  }, {
+    index: 167, 
+    itemName: "Alucart shield", 
+    ramAddr: 0x0A6CF0, 
+    isoAddr: 0x0B3788, 
+    giveItemR4: 0xA7, 
+    giveItemR5: 0x00, 
+    atkVal: 1, 
+    defVal: 0, 
+    elementTxt: "-", 
+    elementHex: 0x0000, 
+    elementOffset: 0x0B3794, 
+    categoryTxt: "Shield", 
+    categoryHex: 0x09, 
+    lockDur: 0, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 6, 
+    shieldSpell: 0x00002000, 
+    mpUse: 0, 
+    hitType: 0x0004, 
+    hitEffect: 0x0001, 
+    menuIcon: 0x0004, 
+    menuPal: 0x0004, 
+    critRate: 1
+  }, {
+    index: 168, 
+    itemName: "Alucart sword", 
+    ramAddr: 0x0A6D24, 
+    isoAddr: 0x0B37BC, 
+    giveItemR4: 0xA8, 
+    giveItemR5: 0x00, 
+    atkVal: 2, 
+    defVal: 0, 
+    elementTxt: "CUT", 
+    elementHex: 0x0040, 
+    elementOffset: 0x0B37C8, 
+    categoryTxt: "Sword", 
+    categoryHex: 0x01, 
+    lockDur: 10, 
+    chainLimit: 0x80, 
+    special: 0x00, 
+    useItem: 0x00, 
+    nmeIFrame: 14, 
+    shieldSpell: 0x00000000, 
+    mpUse: 0, 
+    hitType: 0x0002, 
+    hitEffect: 0x0081, 
+    menuIcon: 0x007B, 
+    menuPal: 0x007B, 
+    critRate: 16
+  }]
+
+  const wearableItems = [{
+    index: 1, 
+    itemName: "Cloth tunic", 
+    ramAddr: 0x0A7718, 
+    isoAddr: 0x0B4300, 
+    giveItemR4: 0x01, 
+    giveItemR5: 0x01, 
+    atkBonus: 0, 
+    defBonus: 1, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4310, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4312, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4314, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4316, 
+    menuIcon: 0x00B9, 
+    menuPal: 0x00B9, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 2, 
+    itemName: "Hide cuirass", 
+    ramAddr: 0x0A7738, 
+    isoAddr: 0x0B4320, 
+    giveItemR4: 0x02, 
+    giveItemR5: 0x02, 
+    atkBonus: 0, 
+    defBonus: 2, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4330, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4332, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4334, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4336, 
+    menuIcon: 0x00BA, 
+    menuPal: 0x00BA, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 3, 
+    itemName: "Bronze cuirass", 
+    ramAddr: 0x0A7758, 
+    isoAddr: 0x0B4340, 
+    giveItemR4: 0x03, 
+    giveItemR5: 0x03, 
+    atkBonus: 0, 
+    defBonus: 3, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4350, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4352, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4354, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4356, 
+    menuIcon: 0x00BB, 
+    menuPal: 0x00BB, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 4, 
+    itemName: "Iron cuirass", 
+    ramAddr: 0x0A7778, 
+    isoAddr: 0x0B4360, 
+    giveItemR4: 0x04, 
+    giveItemR5: 0x04, 
+    atkBonus: 0, 
+    defBonus: 4, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4370, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4372, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4374, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4376, 
+    menuIcon: 0x00BC, 
+    menuPal: 0x00BC, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 5, 
+    itemName: "Steel cuirass", 
+    ramAddr: 0x0A7798, 
+    isoAddr: 0x0B4380, 
+    giveItemR4: 0x05, 
+    giveItemR5: 0x05, 
+    atkBonus: 0, 
+    defBonus: 5, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4390, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4392, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4394, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4396, 
+    menuIcon: 0x00BD, 
+    menuPal: 0x00BD, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 6, 
+    itemName: "Silver plate", 
+    ramAddr: 0x0A77B8, 
+    isoAddr: 0x0B43A0, 
+    giveItemR4: 0x06, 
+    giveItemR5: 0x06, 
+    atkBonus: 0, 
+    defBonus: 6, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B43B0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B43B2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B43B4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B43B6, 
+    menuIcon: 0x00BE, 
+    menuPal: 0x00BE, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 7, 
+    itemName: "Gold plate", 
+    ramAddr: 0x0A77D8, 
+    isoAddr: 0x0B43C0, 
+    giveItemR4: 0x07, 
+    giveItemR5: 0x07, 
+    atkBonus: 0, 
+    defBonus: 8, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4500, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4502, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4504, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4506, 
+    menuIcon: 0x00BF, 
+    menuPal: 0x00BF, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 8, 
+    itemName: "Platinum mail", 
+    ramAddr: 0x0A77F8, 
+    isoAddr: 0x0B4510, 
+    giveItemR4: 0x08, 
+    giveItemR5: 0x08, 
+    atkBonus: 0, 
+    defBonus: 10, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4520, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4522, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4524, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4526, 
+    menuIcon: 0x00C0, 
+    menuPal: 0x00C0, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 9, 
+    itemName: "Diamond plate", 
+    ramAddr: 0x0A7818, 
+    isoAddr: 0x0B4530, 
+    giveItemR4: 0x09, 
+    giveItemR5: 0x09, 
+    atkBonus: 0, 
+    defBonus: 15, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4540, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4542, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4544, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4546, 
+    menuIcon: 0x00C1, 
+    menuPal: 0x00C1, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 10, 
+    itemName: "Fire mail", 
+    ramAddr: 0x0A7838, 
+    isoAddr: 0x0B4550, 
+    giveItemR4: 0x0A, 
+    giveItemR5: 0x10, 
+    atkBonus: 0, 
+    defBonus: 4, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4560, 
+    elemResist: "FIR", 
+    elemResistHex: 0x8000, 
+    elemResistOffset: 0x0B4562, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4564, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4566, 
+    menuIcon: 0x00C2, 
+    menuPal: 0x00C2, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 11, 
+    itemName: "Lightning mail", 
+    ramAddr: 0x0A7858, 
+    isoAddr: 0x0B4570, 
+    giveItemR4: 0x0B, 
+    giveItemR5: 0x11, 
+    atkBonus: 0, 
+    defBonus: 8, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4580, 
+    elemResist: "LIT", 
+    elemResistHex: 0x4000, 
+    elemResistOffset: 0x0B4582, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4584, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4586, 
+    menuIcon: 0x00C3, 
+    menuPal: 0x00C3, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 12, 
+    itemName: "Ice mail", 
+    ramAddr: 0x0A7878, 
+    isoAddr: 0x0B4590, 
+    giveItemR4: 0x0C, 
+    giveItemR5: 0x12, 
+    atkBonus: 0, 
+    defBonus: 8, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B45A0, 
+    elemResist: "ICE", 
+    elemResistHex: 0x2000, 
+    elemResistOffset: 0x0B45A2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B45A4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B45A6, 
+    menuIcon: 0x00C4, 
+    menuPal: 0x00C4, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 13, 
+    itemName: "Mirror cuirass", 
+    ramAddr: 0x0A7898, 
+    isoAddr: 0x0B45B0, 
+    giveItemR4: 0x0D, 
+    giveItemR5: 0x13, 
+    atkBonus: 0, 
+    defBonus: 3, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B45C0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B45C2, 
+    elemGuard: "STN", 
+    elemGuardHex: 0x0200, 
+    elemGuardOffset: 0x0B45C4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B45C6, 
+    menuIcon: 0x00C5, 
+    menuPal: 0x00C5, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 14, 
+    itemName: "Spike Breaker", 
+    ramAddr: 0x0A78B8, 
+    isoAddr: 0x0B45D0, 
+    giveItemR4: 0x0E, 
+    giveItemR5: 0x14, 
+    atkBonus: 0, 
+    defBonus: 12, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B45E0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B45E2, 
+    elemGuard: "PCT", 
+    elemGuardHex: 0x0010, 
+    elemGuardOffset: 0x0B45E4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B45E6, 
+    menuIcon: 0x00C6, 
+    menuPal: 0x00C6, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 15, 
+    itemName: "Alucard mail", 
+    ramAddr: 0x0A78D8, 
+    isoAddr: 0x0B45F0, 
+    giveItemR4: 0x0F, 
+    giveItemR5: 0x15, 
+    atkBonus: 0, 
+    defBonus: 20, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4600, 
+    elemResist: "FIR+LIT+ICE", 
+    elemResistHex: 0xE000, 
+    elemResistOffset: 0x0B4602, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4604, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4606, 
+    menuIcon: 0x00C7, 
+    menuPal: 0x00C7, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 16, 
+    itemName: "Dark armor", 
+    ramAddr: 0x0A78F8, 
+    isoAddr: 0x0B4610, 
+    giveItemR4: 0x10, 
+    giveItemR5: 0x16, 
+    atkBonus: 0, 
+    defBonus: 8, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    "elemWeak": "HLY", 
+    elemWeakHex: 0x1000, 
+    elemWeakOffset: 0x0B4620, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4622, 
+    elemGuard: "DRK", 
+    elemGuardHex: 0x0800, 
+    elemGuardOffset: 0x0B4624, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4626, 
+    menuIcon: 0x00C8, 
+    menuPal: 0x00C8, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 17, 
+    itemName: "Healing mail", 
+    ramAddr: 0x0A7918, 
+    isoAddr: 0x0B4630, 
+    giveItemR4: 0x11, 
+    giveItemR5: 0x17, 
+    atkBonus: 0, 
+    defBonus: 4, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4640, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4642, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4644, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4646, 
+    menuIcon: 0x00C9, 
+    menuPal: 0x00C9, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 18, 
+    itemName: "Holy mail", 
+    ramAddr: 0x0A7938, 
+    isoAddr: 0x0B4650, 
+    giveItemR4: 0x12, 
+    giveItemR5: 0x18, 
+    atkBonus: 0, 
+    defBonus: 11, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4660, 
+    elemResist: "HLY", 
+    elemResistHex: 0x1000, 
+    elemResistOffset: 0x0B4662, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4664, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4666, 
+    menuIcon: 0x00CA, 
+    menuPal: 0x00CA, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 19, 
+    itemName: "Walk armor", 
+    ramAddr: 0x0A7958, 
+    isoAddr: 0x0B4670, 
+    giveItemR4: 0x13, 
+    giveItemR5: 0x19, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: -5, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4680, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4682, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4684, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4686, 
+    menuIcon: 0x00CB, 
+    menuPal: 0x00CB, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 20, 
+    itemName: "Brilliant mail", 
+    ramAddr: 0x0A7978, 
+    isoAddr: 0x0B4690, 
+    giveItemR4: 0x14, 
+    giveItemR5: 0x20, 
+    atkBonus: 0, 
+    defBonus: 12, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 1, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B46A0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B46A2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B46A4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B46A6, 
+    menuIcon: 0x00CC, 
+    menuPal: 0x00CC, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 21, 
+    itemName: "Mojo mail", 
+    ramAddr: 0x0A7998, 
+    isoAddr: 0x0B46B0, 
+    giveItemR4: 0x15, 
+    giveItemR5: 0x21, 
+    atkBonus: 0, 
+    defBonus: 15, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 1, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B46C0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B46C2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B46C4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B46C6, 
+    menuIcon: 0x00CD, 
+    menuPal: 0x00CD, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 22, 
+    itemName: "Fury plate", 
+    ramAddr: 0x0A79B8, 
+    isoAddr: 0x0B46D0, 
+    giveItemR4: 0x16, 
+    giveItemR5: 0x22, 
+    atkBonus: 1, 
+    defBonus: 13, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B46E0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B46E2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B46E4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B46E6, 
+    menuIcon: 0x00CE, 
+    menuPal: 0x00CE, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 23, 
+    itemName: "Dracula tunic", 
+    ramAddr: 0x0A79D8, 
+    isoAddr: 0x0B46F0, 
+    giveItemR4: 0x17, 
+    giveItemR5: 0x23, 
+    atkBonus: 2, 
+    defBonus: 14, 
+    strBonus: 1, 
+    intBonus: 1, 
+    conBonus: 1, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4700, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4702, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4704, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4706, 
+    menuIcon: 0x00CF, 
+    menuPal: 0x00CF, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 24, 
+    itemName: "God\'s Garb", 
+    ramAddr: 0x0A79F8, 
+    isoAddr: 0x0B4710, 
+    giveItemR4: 0x18, 
+    giveItemR5: 0x24, 
+    atkBonus: 0, 
+    defBonus: 25, 
+    strBonus: 0, 
+    intBonus: 2, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4720, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4722, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4724, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4726, 
+    menuIcon: 0x00D0, 
+    menuPal: 0x00D0, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 25, 
+    itemName: "Axe Lord armor", 
+    ramAddr: 0x0A7A18, 
+    isoAddr: 0x0B4730, 
+    giveItemR4: 0x19, 
+    giveItemR5: 0x25, 
+    atkBonus: 2, 
+    defBonus: 15, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4740, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4742, 
+    elemGuard: "STN", 
+    elemGuardHex: 0x0200, 
+    elemGuardOffset: 0x0B4744, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4746, 
+    menuIcon: 0x00D1, 
+    menuPal: 0x00D1, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }, {
+    index: 26, 
+    itemName: "Sunglasses", 
+    ramAddr: 0x0A7A38, 
+    isoAddr: 0x0B4770, 
+    giveItemR4: 0x1B, 
+    giveItemR5: 0x26, 
+    atkBonus: 0, 
+    defBonus: 1, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: -1, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4780, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4782, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4784, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4786, 
+    menuIcon: 0x00A0, 
+    menuPal: 0x00A0, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 27, 
+    itemName: "Ballroom mask", 
+    ramAddr: 0x0A7A58, 
+    isoAddr: 0x0B4790, 
+    giveItemR4: 0x1C, 
+    giveItemR5: 0x27, 
+    atkBonus: 0, 
+    defBonus: 2, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B47A0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B47A2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B47A4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B47A6, 
+    menuIcon: 0x00A1, 
+    menuPal: 0x00A1, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 28, 
+    itemName: "Bandanna", 
+    ramAddr: 0x0A7A78, 
+    isoAddr: 0x0B47B0, 
+    giveItemR4: 0x1D, 
+    giveItemR5: 0x28, 
+    atkBonus: 0, 
+    defBonus: 2, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 1, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B47C0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B47C2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B47C4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B47C6, 
+    menuIcon: 0x00A2, 
+    menuPal: 0x00A2, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 29, 
+    itemName: "Felt hat", 
+    ramAddr: 0x0A7A98, 
+    isoAddr: 0x0B47D0, 
+    giveItemR4: 0x1E, 
+    giveItemR5: 0x29, 
+    atkBonus: 0, 
+    defBonus: 1, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 1, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B47E0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B47E2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B47E4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B47E6, 
+    menuIcon: 0x00A3, 
+    menuPal: 0x00A3, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 30, 
+    itemName: "Velvet hat", 
+    ramAddr: 0x0A7AB8, 
+    isoAddr: 0x0B47F0, 
+    giveItemR4: 0x1F, 
+    giveItemR5: 0x30, 
+    atkBonus: 0, 
+    defBonus: 1, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 2, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4800, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4802, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4804, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4806, 
+    menuIcon: 0x00A4, 
+    menuPal: 0x00A4, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 31, 
+    itemName: "Goggles", 
+    ramAddr: 0x0A7AD8, 
+    isoAddr: 0x0B4810, 
+    giveItemR4: 0x20, 
+    giveItemR5: 0x31, 
+    atkBonus: 0, 
+    defBonus: 1, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 4, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4820, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4822, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4824, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4826, 
+    menuIcon: 0x00A5, 
+    menuPal: 0x00A5, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 32, 
+    itemName: "Leather hat", 
+    ramAddr: 0x0A7AF8, 
+    isoAddr: 0x0B4830, 
+    giveItemR4: 0x21, 
+    giveItemR5: 0x32, 
+    atkBonus: 0, 
+    defBonus: 2, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4840, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4842, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4844, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4846, 
+    menuIcon: 0x00A6, 
+    menuPal: 0x00A6, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 33, 
+    itemName: "Holy glasses", 
+    ramAddr: 0x0A7B18, 
+    isoAddr: 0x0B4850, 
+    giveItemR4: 0x22, 
+    giveItemR5: 0x33, 
+    atkBonus: 0, 
+    defBonus: 1, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 5, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4860, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4862, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4864, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4866, 
+    menuIcon: 0x00A7, 
+    menuPal: 0x00A7, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 34, 
+    itemName: "Steel helm", 
+    ramAddr: 0x0A7B38, 
+    isoAddr: 0x0B4870, 
+    giveItemR4: 0x23, 
+    giveItemR5: 0x34, 
+    atkBonus: 0, 
+    defBonus: 3, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 1, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4880, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4882, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4884, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4886, 
+    menuIcon: 0x00A8, 
+    menuPal: 0x00A8, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 35, 
+    itemName: "Stone mask", 
+    ramAddr: 0x0A7B58, 
+    isoAddr: 0x0B4890, 
+    giveItemR4: 0x24, 
+    giveItemR5: 0x35, 
+    atkBonus: 0, 
+    defBonus: 2, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 4, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B48A0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B48A2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B48A4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B48A6, 
+    menuIcon: 0x00A9, 
+    menuPal: 0x00A9, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 36, 
+    itemName: "Circlet", 
+    ramAddr: 0x0A7B78, 
+    isoAddr: 0x0B48B0, 
+    giveItemR4: 0x25, 
+    giveItemR5: 0x36, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 5, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B48C0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B48C2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B48C4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B48C6, 
+    menuIcon: 0x00AA, 
+    menuPal: 0x00AA, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 37, 
+    itemName: "Gold circlet", 
+    ramAddr: 0x0A7B98, 
+    isoAddr: 0x0B48D0, 
+    giveItemR4: 0x26, 
+    giveItemR5: 0x37, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 6, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B48E0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B48E2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B48E4, 
+    elemAbsorb: "HLY", 
+    elemAbsorbHex: 0x1000, 
+    elemAbsorbOffset: 0x0B48E6, 
+    menuIcon: 0x00AB, 
+    menuPal: 0x00AB, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 38, 
+    itemName: "Ruby circlet", 
+    ramAddr: 0x0A7BB8, 
+    isoAddr: 0x0B48F0, 
+    giveItemR4: 0x27, 
+    giveItemR5: 0x38, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 10, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4900, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4902, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4904, 
+    elemAbsorb: "FIR", 
+    elemAbsorbHex: 0x8000, 
+    elemAbsorbOffset: 0x0B4906, 
+    menuIcon: 0x00AC, 
+    menuPal: 0x0093, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 39, 
+    itemName: "Opal circlet", 
+    ramAddr: 0x0A7BD8, 
+    isoAddr: 0x0B4910, 
+    giveItemR4: 0x28, 
+    giveItemR5: 0x39, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 10, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4920, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4922, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4924, 
+    elemAbsorb: "ICE", 
+    elemAbsorbHex: 0x2000, 
+    elemAbsorbOffset: 0x0B4926, 
+    menuIcon: 0x00AC, 
+    menuPal: 0x0097, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 40, 
+    itemName: "Topaz circlet", 
+    ramAddr: 0x0A7BF8, 
+    isoAddr: 0x0B4930, 
+    giveItemR4: 0x29, 
+    giveItemR5: 0x40, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 10, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4940, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4942, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4944, 
+    elemAbsorb: "PSN", 
+    elemAbsorbHex: 0x0080, 
+    elemAbsorbOffset: 0x0B4946, 
+    menuIcon: 0x00AC, 
+    menuPal: 0x0095, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 41, 
+    itemName: "Beryl circlet", 
+    ramAddr: 0x0A7C18, 
+    isoAddr: 0x0B4950, 
+    giveItemR4: 0x2A, 
+    giveItemR5: 0x41, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 10, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4960, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4962, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4964, 
+    elemAbsorb: "LIT", 
+    elemAbsorbHex: 0x4000, 
+    elemAbsorbOffset: 0x0B4966, 
+    menuIcon: 0x00AD, 
+    menuPal: 0x0096, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 42, 
+    itemName: "Cat-eye circl.", 
+    ramAddr: 0x0A7C38, 
+    isoAddr: 0x0B4970, 
+    giveItemR4: 0x2B, 
+    giveItemR5: 0x42, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 8, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4980, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4982, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4984, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4986, 
+    menuIcon: 0x00AD, 
+    menuPal: 0x0095, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 43, 
+    itemName: "Coral circlet", 
+    ramAddr: 0x0A7C58, 
+    isoAddr: 0x0B4990, 
+    giveItemR4: 0x2C, 
+    giveItemR5: 0x43, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 7, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B49A0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B49A2, 
+    elemGuard: "CUR", 
+    elemGuardHex: 0x0100, 
+    elemGuardOffset: 0x0B49A4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B49A6, 
+    menuIcon: 0x00AA, 
+    menuPal: 0x00AC, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 44, 
+    itemName: "Dragon helm", 
+    ramAddr: 0x0A7C78, 
+    isoAddr: 0x0B49B0, 
+    giveItemR4: 0x2D, 
+    giveItemR5: 0x44, 
+    atkBonus: 0, 
+    defBonus: 3, 
+    strBonus: 4, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B49C0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B49C2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B49C4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B49C6, 
+    menuIcon: 0x00B2, 
+    menuPal: 0x00B2, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 45, 
+    itemName: "Silver crown", 
+    ramAddr: 0x0A7C98, 
+    isoAddr: 0x0B49D0, 
+    giveItemR4: 0x2E, 
+    giveItemR5: 0x45, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 12, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B49E0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B49E2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B49E4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B49E6, 
+    menuIcon: 0x00B3, 
+    menuPal: 0x00B3, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 46, 
+    itemName: "Wizard hat", 
+    ramAddr: 0x0A7CB8, 
+    isoAddr: 0x0B49F0, 
+    giveItemR4: 0x2F, 
+    giveItemR5: 0x46, 
+    atkBonus: 0, 
+    defBonus: 5, 
+    strBonus: 1, 
+    intBonus: 1, 
+    conBonus: 7, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4A00, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4A02, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4A04, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4A06, 
+    menuIcon: 0x00B4, 
+    menuPal: 0x00B4, 
+    equipTypeTxt: "Helm", 
+    equipTypeHex: 0x0000
+  }, {
+    index: 47, 
+    itemName: "Cloth cape", 
+    ramAddr: 0x0A7CD8, 
+    isoAddr: 0x0B4A30, 
+    giveItemR4: 0x31, 
+    giveItemR5: 0x47, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 1, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4A40, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4A42, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4A44, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4A46, 
+    menuIcon: 0x00D8, 
+    menuPal: 0x00D8, 
+    equipTypeTxt: "Cape", 
+    equipTypeHex: 0x0002
+  }, {
+    index: 48, 
+    itemName: "Reverse cloak", 
+    ramAddr: 0x0A7CF8, 
+    isoAddr: 0x0B4A50, 
+    giveItemR4: 0x32, 
+    giveItemR5: 0x48, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 3, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4A60, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4A62, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4A64, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4A66, 
+    menuIcon: 0x00D9, 
+    menuPal: 0x00D9, 
+    equipTypeTxt: "Cape", 
+    equipTypeHex: 0x0002
+  }, {
+    index: 49, 
+    itemName: "Elven cloak", 
+    ramAddr: 0x0A7D18, 
+    isoAddr: 0x0B4A70, 
+    giveItemR4: 0x33, 
+    giveItemR5: 0x49, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 5, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4A80, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4A82, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4A84, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4A86, 
+    menuIcon: 0x00DA, 
+    menuPal: 0x00DA, 
+    equipTypeTxt: "Cape", 
+    equipTypeHex: 0x0002
+  }, {
+    index: 50, 
+    itemName: "Crystal cloak", 
+    ramAddr: 0x0A7D38, 
+    isoAddr: 0x0B4A90, 
+    giveItemR4: 0x34, 
+    giveItemR5: 0x50, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 6, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4AA0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4AA2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4AA4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4AA6, 
+    menuIcon: 0x00DB, 
+    menuPal: 0x00DB, 
+    equipTypeTxt: "Cape", 
+    equipTypeHex: 0x0002
+  }, {
+    index: 51, 
+    itemName: "Royal cloak", 
+    ramAddr: 0x0A7D58, 
+    isoAddr: 0x0B4AB0, 
+    giveItemR4: 0x35, 
+    giveItemR5: 0x51, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 8, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4AC0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4AC2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4AC4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4AC6, 
+    menuIcon: 0x00DC, 
+    menuPal: 0x00DC, 
+    equipTypeTxt: "Cape", 
+    equipTypeHex: 0x0002
+  }, {
+    index: 52, 
+    itemName: "Blood cloak", 
+    ramAddr: 0x0A7D78, 
+    isoAddr: 0x0B4AD0, 
+    giveItemR4: 0x36, 
+    giveItemR5: 0x52, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 4, 
+    conBonus: 0, 
+    lckBonus: -1, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4AE0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4AE2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4AE4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4AE6, 
+    menuIcon: 0x00DD, 
+    menuPal: 0x00DD, 
+    equipTypeTxt: "Cape", 
+    equipTypeHex: 0x0002
+  }, {
+    index: 53, 
+    itemName: "Joseph\'s cloak", 
+    ramAddr: 0x0A7D98, 
+    isoAddr: 0x0B4AF0, 
+    giveItemR4: 0x37, 
+    giveItemR5: 0x53, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 9, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4B00, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4B02, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4B04, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4B06, 
+    menuIcon: 0x00DE, 
+    menuPal: 0x00DE, 
+    equipTypeTxt: "Cape", 
+    equipTypeHex: 0x0002
+  }, {
+    index: 54, 
+    itemName: "Twilight cloak", 
+    ramAddr: 0x0A7DB8, 
+    isoAddr: 0x0B4B10, 
+    giveItemR4: 0x38, 
+    giveItemR5: 0x54, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 10, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4B20, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4B22, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4B24, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4B26, 
+    menuIcon: 0x00DF, 
+    menuPal: 0x00DF, 
+    equipTypeTxt: "Cape", 
+    equipTypeHex: 0x0002
+  }, {
+    index: 55, 
+    itemName: "Moonstone", 
+    ramAddr: 0x0A7DD8, 
+    isoAddr: 0x0B4B50, 
+    giveItemR4: 0x3A, 
+    giveItemR5: 0x55, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4B60, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4B62, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4B64, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4B66, 
+    menuIcon: 0x00E0, 
+    menuPal: 0x00E0, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 56, 
+    itemName: "Sunstone", 
+    ramAddr: 0x0A7DF8, 
+    isoAddr: 0x0B4B70, 
+    giveItemR4: 0x3B, 
+    giveItemR5: 0x56, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4B80, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4B82, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4B84, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4B86, 
+    menuIcon: 0x00E1, 
+    menuPal: 0x00E1, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 57, 
+    itemName: "Bloodstone", 
+    ramAddr: 0x0A7E18, 
+    isoAddr: 0x0B4B90, 
+    giveItemR4: 0x3C, 
+    giveItemR5: 0x57, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4BA0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4BA2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4BA4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4BA6, 
+    menuIcon: 0x00E2, 
+    menuPal: 0x00E2, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 58, 
+    itemName: "Staurolite", 
+    ramAddr: 0x0A7E38, 
+    isoAddr: 0x0B4BB0, 
+    giveItemR4: 0x3D, 
+    giveItemR5: 0x58, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4BC0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4BC2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4BC4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4BC6, 
+    menuIcon: 0x00E3, 
+    menuPal: 0x00E3, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 59, 
+    itemName: "Ring of Pales", 
+    ramAddr: 0x0A7E58, 
+    isoAddr: 0x0B4BD0, 
+    giveItemR4: 0x3E, 
+    giveItemR5: 0x59, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 1, 
+    intBonus: 1, 
+    conBonus: 1, 
+    lckBonus: 1, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4BE0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4BE2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4BE4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4BE6, 
+    menuIcon: 0x00E4, 
+    menuPal: 0x00E4, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 60, 
+    itemName: "Zircon", 
+    ramAddr: 0x0A7E78, 
+    isoAddr: 0x0B4BF0, 
+    giveItemR4: 0x3F, 
+    giveItemR5: 0x60, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4C00, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4C02, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4C04, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4C06, 
+    menuIcon: 0x00E6, 
+    menuPal: 0x0096, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 61, 
+    itemName: "Aquamarine", 
+    ramAddr: 0x0A7E98, 
+    isoAddr: 0x0B4C10, 
+    giveItemR4: 0x40, 
+    giveItemR5: 0x61, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4C20, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4C22, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4C24, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4C26, 
+    menuIcon: 0x00E6, 
+    menuPal: 0x0097, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 62, 
+    itemName: "Turquoise", 
+    ramAddr: 0x0A7EB8, 
+    isoAddr: 0x0B4C30, 
+    giveItemR4: 0x41, 
+    giveItemR5: 0x62, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4C40, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4C42, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4C44, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4C46, 
+    menuIcon: 0x00E6, 
+    menuPal: 0x0099, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 63, 
+    itemName: "Onyx", 
+    ramAddr: 0x0A7ED8, 
+    isoAddr: 0x0B4C50, 
+    giveItemR4: 0x42, 
+    giveItemR5: 0x63, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4C60, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4C62, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4C64, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4C66, 
+    menuIcon: 0x00E6, 
+    menuPal: 0x009A, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 64, 
+    itemName: "Garnet", 
+    ramAddr: 0x0A7EF8, 
+    isoAddr: 0x0B4C70, 
+    giveItemR4: 0x43, 
+    giveItemR5: 0x64, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4C80, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4C82, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4C84, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4C86, 
+    menuIcon: 0x00E6, 
+    menuPal: 0x0094, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 65, 
+    itemName: "Opal", 
+    ramAddr: 0x0A7F18, 
+    isoAddr: 0x0B4C90, 
+    giveItemR4: 0x44, 
+    giveItemR5: 0x65, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4CA0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4CA2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4CA4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4CA6, 
+    menuIcon: 0x00E6, 
+    menuPal: 0x0095, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 66, 
+    itemName: "Diamond", 
+    ramAddr: 0x0A7F38, 
+    isoAddr: 0x0B4CB0, 
+    giveItemR4: 0x45, 
+    giveItemR5: 0x66, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4CC0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4CC2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4CC4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4CC6, 
+    menuIcon: 0x00E5, 
+    menuPal: 0x00E5, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 67, 
+    itemName: "Lapis lazuli", 
+    ramAddr: 0x0A7F58, 
+    isoAddr: 0x0B4CD0, 
+    giveItemR4: 0x46, 
+    giveItemR5: 0x67, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 20, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4CE0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4CE2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4CE4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4CE6, 
+    menuIcon: 0x00E6, 
+    menuPal: 0x0091, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 68, 
+    itemName: "Ring of Ares", 
+    ramAddr: 0x0A7F78, 
+    isoAddr: 0x0B4CF0, 
+    giveItemR4: 0x47, 
+    giveItemR5: 0x68, 
+    atkBonus: 0, 
+    defBonus: -24, 
+    strBonus: 24, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4E30, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4E32, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4E34, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4E36, 
+    menuIcon: 0x00ED, 
+    menuPal: 0x00ED, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 69, 
+    itemName: "Gold Ring", 
+    ramAddr: 0x0A7F98, 
+    isoAddr: 0x0B4E40, 
+    giveItemR4: 0x48, 
+    giveItemR5: 0x69, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4E50, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4E52, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4E54, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4E56, 
+    menuIcon: 0x00EF, 
+    menuPal: 0x00EF, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 70, 
+    itemName: "Silver Ring", 
+    ramAddr: 0x0A7FB8, 
+    isoAddr: 0x0B4E60, 
+    giveItemR4: 0x49, 
+    giveItemR5: 0x70, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4E70, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4E72, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4E74, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4E76, 
+    menuIcon: 0x00EE, 
+    menuPal: 0x00EE, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 71, 
+    itemName: "Ring of Varda", 
+    ramAddr: 0x0A7FD8, 
+    isoAddr: 0x0B4E80, 
+    giveItemR4: 0x4A, 
+    giveItemR5: 0x71, 
+    atkBonus: 10, 
+    defBonus: 10, 
+    strBonus: 30, 
+    intBonus: 10, 
+    conBonus: 30, 
+    lckBonus: 5, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4E90, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4E92, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4E94, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4E96, 
+    menuIcon: 0x00F0, 
+    menuPal: 0x00F0, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 72, 
+    itemName: "Ring of Arcana", 
+    ramAddr: 0x0A7FF8, 
+    isoAddr: 0x0B4EA0, 
+    giveItemR4: 0x4B, 
+    giveItemR5: 0x72, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4EB0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4EB2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4EB4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4EB6, 
+    menuIcon: 0x00F1, 
+    menuPal: 0x0095, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 73, 
+    itemName: "Mystic pendant", 
+    ramAddr: 0x0A8018, 
+    isoAddr: 0x0B4EC0, 
+    giveItemR4: 0x4C, 
+    giveItemR5: 0x73, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4ED0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4ED2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4ED4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4ED6, 
+    menuIcon: 0x00F2, 
+    menuPal: 0x00F2, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 74, 
+    itemName: "Heart broach", 
+    ramAddr: 0x0A8038, 
+    isoAddr: 0x0B4EE0, 
+    giveItemR4: 0x4D, 
+    giveItemR5: 0x74, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4EF0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4EF2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4EF4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4EF6, 
+    menuIcon: 0x00F3, 
+    menuPal: 0x00F3, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 75, 
+    itemName: "Necklace of J", 
+    ramAddr: 0x0A8058, 
+    isoAddr: 0x0B4F00, 
+    giveItemR4: 0x4E, 
+    giveItemR5: 0x75, 
+    atkBonus: 0, 
+    defBonus: 5, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4F10, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4F12, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4F14, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4F16, 
+    menuIcon: 0x00F4, 
+    menuPal: 0x00F4, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 76, 
+    itemName: "Gauntlet", 
+    ramAddr: 0x0A8078, 
+    isoAddr: 0x0B4F20, 
+    giveItemR4: 0x4F, 
+    giveItemR5: 0x76, 
+    atkBonus: 5, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4F30, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4F32, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4F34, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4F36, 
+    menuIcon: 0x00F5, 
+    menuPal: 0x00F5, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 77, 
+    itemName: "Ankh of Life", 
+    ramAddr: 0x0A8098, 
+    isoAddr: 0x0B4F40, 
+    giveItemR4: 0x50, 
+    giveItemR5: 0x77, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4F50, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4F52, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4F54, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4F56, 
+    menuIcon: 0x00F6, 
+    menuPal: 0x00F6, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 78, 
+    itemName: "Ring of Feanor", 
+    ramAddr: 0x0A80B8, 
+    isoAddr: 0x0B4F60, 
+    giveItemR4: 0x51, 
+    giveItemR5: 0x78, 
+    atkBonus: 1, 
+    defBonus: 1, 
+    strBonus: 7, 
+    intBonus: 7, 
+    conBonus: 7, 
+    lckBonus: 2, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4F70, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4F72, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4F74, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4F76, 
+    menuIcon: 0x00F7, 
+    menuPal: 0x00F7, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 79, 
+    itemName: "Medal", 
+    ramAddr: 0x0A80D8, 
+    isoAddr: 0x0B4F80, 
+    giveItemR4: 0x52, 
+    giveItemR5: 0x79, 
+    atkBonus: 1, 
+    defBonus: 1, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4F90, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4F92, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4F94, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4F96, 
+    menuIcon: 0x00F8, 
+    menuPal: 0x00F8, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 80, 
+    itemName: "Talisman", 
+    ramAddr: 0x0A80F8, 
+    isoAddr: 0x0B4FA0, 
+    giveItemR4: 0x53, 
+    giveItemR5: 0x80, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4FB0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4FB2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4FB4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4FB6, 
+    menuIcon: 0x00F9, 
+    menuPal: 0x00F9, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 81, 
+    itemName: "Duplicator", 
+    ramAddr: 0x0A8118, 
+    isoAddr: 0x0B4FC0, 
+    giveItemR4: 0x54, 
+    giveItemR5: 0x81, 
+    atkBonus: 0, 
+    defBonus: -20, 
+    strBonus: 0, 
+    intBonus: -10, 
+    conBonus: -10, 
+    lckBonus: -10, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4FD0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4FD2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4FD4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4FD6, 
+    menuIcon: 0x00FA, 
+    menuPal: 0x00FA, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 82, 
+    itemName: "King\'s stone", 
+    ramAddr: 0x0A8138, 
+    isoAddr: 0x0B4FE0, 
+    giveItemR4: 0x55, 
+    giveItemR5: 0x82, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 10, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B4FF0, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B4FF2, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B4FF4, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B4FF6, 
+    menuIcon: 0x00FB, 
+    menuPal: 0x00FB, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 83, 
+    itemName: "Covenant stone", 
+    ramAddr: 0x0A8158, 
+    isoAddr: 0x0B5000, 
+    giveItemR4: 0x56, 
+    giveItemR5: 0x83, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 15, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B5010, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B5012, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B5014, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B5016, 
+    menuIcon: 0x00FC, 
+    menuPal: 0x00FC, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 84, 
+    itemName: "Nauglamir", 
+    ramAddr: 0x0A8178, 
+    isoAddr: 0x0B5020, 
+    giveItemR4: 0x57, 
+    giveItemR5: 0x84, 
+    atkBonus: 0, 
+    defBonus: 15, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B5030, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B5032, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B5034, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B5036, 
+    menuIcon: 0x00FD, 
+    menuPal: 0x00FD, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 85, 
+    itemName: "Secret boots", 
+    ramAddr: 0x0A8198, 
+    isoAddr: 0x0B5040, 
+    giveItemR4: 0x58, 
+    giveItemR5: 0x85, 
+    atkBonus: 0, 
+    defBonus: 0, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B5050, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B5052, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B5054, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B5056, 
+    menuIcon: 0x00FE, 
+    menuPal: 0x00FE, 
+    equipTypeTxt: "Accessory", 
+    equipTypeHex: 0x0003
+  }, {
+    index: 86, 
+    itemName: "Alucart mail", 
+    ramAddr: 0x0A81B8, 
+    isoAddr: 0x0B5060, 
+    giveItemR4: 0x59, 
+    giveItemR5: 0x86, 
+    atkBonus: 0, 
+    defBonus: 2, 
+    strBonus: 0, 
+    intBonus: 0, 
+    conBonus: 0, 
+    lckBonus: 0, 
+    elemWeakHex: 0x0000, 
+    elemWeakOffset: 0x0B5070, 
+    elemResistHex: 0x0000, 
+    elemResistOffset: 0x0B5072, 
+    elemGuardHex: 0x0000, 
+    elemGuardOffset: 0x0B5074, 
+    elemAbsorbHex: 0x0000, 
+    elemAbsorbOffset: 0x0B5076, 
+    menuIcon: 0x00C7, 
+    menuPal: 0x00C7, 
+    equipTypeTxt: "Armor", 
+    equipTypeHex: 0x0001
+  }]
+
+  const exports = {
+    handItems: handItems,
+    wearableItems: wearableItems
+  }
+
+  if (self) {
+    self.sotnRando = Object.assign(self.sotnRando || {}, {
+      equipment: exports,
+    })
+  } else {
+    module.exports = exports
+  }
+})(typeof(self) !== 'undefined' ? self : null)
