+++ src/items.js
@@ -9336,7 +9336,8 @@
       entities: [ 0x3698, 0x4518 ],
     }],
   }]
-
+  
+  
   const exports = items
 
   if (self) {
